+++ CSMI.cs
@@ -18,15 +18,18 @@ using System.Collections.Concurrent;
 
 namespace CSMI
 {
-    public class MI : IDisposable
+    public class Constants
     {
-        const double LOG_BASE = 2.0;
-        const int MAX_YDIM = 65535;
-        const int GRIDSIZE = 1024;
-        Device dev;
-        Context context;
+        public const double LOG_BASE = 2.0;
+    }
 
-        public MI()
+    // TODO: Rename this to something better
+    public class ILGPUInitializer : IDisposable
+    {
+        public Device dev;
+        public Context context;
+
+        public ILGPUInitializer()
         {
             this.context = Context.Create(builder => builder.AllAccelerators().EnableAlgorithms());
             this.dev = this.context.GetPreferredDevice(preferCPU: false);
@@ -50,6 +53,29 @@ namespace CSMI
         public void Dispose()
         {
             this.context.Dispose();
+        }
+    }
+
+    public class CSMI : IDisposable
+    {
+        public double LOG_BASE = 2.0;
+        public const int MAX_YDIM = 65535;
+        public const int GRIDSIZE = 1024;
+
+        public ILGPUInitializer ILGPUInitializer { get; private set; }
+        public Entropy entropy { get; private set; }
+        public MutualInformation mi { get; private set; }
+
+        public CSMI()
+        {
+            ILGPUInitializer = new ILGPUInitializer();
+            entropy = new Entropy(ILGPUInitializer);
+            mi = new MutualInformation(ILGPUInitializer);
+        }
+
+        public void Dispose()
+        {
+            ILGPUInitializer.Dispose();
         }
 
         public double[] refactorArray(double[] arr)
@@ -120,1708 +146,93 @@ namespace CSMI
             return refactoredarr;
         }
 
-        public static long szudzikPair(long x, long y)
-        {
-            return (x >= y ? (x * x) + x + y : (y * y) + x);
-        }
-
-        public double calculateEntropy(double[] dataVector)
-        {
-            using Accelerator accelerate = this.dev.CreateAccelerator(this.context);
-            using var MVBuffer = accelerate.Allocate1D<double>(new Index1D(dataVector.GetLength(0)));
-            using var MVNormBuffer = accelerate.Allocate1D<double>(new Index1D(dataVector.GetLength(0)));
-            //var FreqBuffer = accelerate.Allocate1D<double>(new Index1D(dataVector.GetLength(0)));
-            MVBuffer.CopyFromCPU(dataVector);
-            using var MaxVal = accelerate.Allocate1D<int>(new Index1D(1));
-            using var MinVal = accelerate.Allocate1D<int>(new Index1D(1));
-            using var EntropyBuffer = accelerate.Allocate1D<double>(new Index1D(1));
-
-            var GetMaxValKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>
-            >(GetMaxMinValKernal);
-
-            var InitMaxMinKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>
-            >(InitMaxMinKernel);
-
-            var BuildFreqKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>
-            >(BuildFreqAdjKernel);
-
-            var CalcEntropyKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>,
-                int
-            >(CalcEntropyKernel);
-
-            var normalizeArrayKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>
-            >(normalizeArrayKernel);
-
-            InitMaxMinKern(MaxVal.Extent.ToIntIndex(), MVBuffer.View, MaxVal.View, MinVal.View);
-            GetMaxValKern(MVBuffer.Extent.ToIntIndex(), MVBuffer.View, MaxVal.View, MinVal.View);
-            normalizeArrayKern(MVBuffer.Extent.ToIntIndex(), MVBuffer.View, MVNormBuffer.View, MinVal);
-            using var FreqBuffer = accelerate.Allocate1D<double>(
-                new Index1D(MaxVal.GetAsArray1D()[0] - MinVal.GetAsArray1D()[0] + 1)
-            );
-            //setBuffToValueKern(FreqBuffer.Extent.ToIntIndex(), FreqBuffer.View, 0.0);
-            // if(dataVector.GetLength(0) > MAX_YDIM){
-            //     int zdim = (int)Math.Floor((double)MVBuffer.Extent.ToIntIndex().X/GRIDSIZE) + 1;
-            //     int ydim = GRIDSIZE;
-            //     Console.WriteLine("YDIM AND ZDIM");
-            //     Console.WriteLine(ydim);
-            //     Console.WriteLine(zdim);
-            //     LargeBuildFreqKern(new Index3D(MVBuffer.Extent.ToIntIndex().X,ydim, zdim), MVBuffer.View, FreqBuffer.View, GRIDSIZE, dataVector.GetLength(0));
-            // }
-            // else{
-            //BuildFreqKern(new Index2D(MVBuffer.Extent.ToIntIndex().X,MVBuffer.Extent.ToIntIndex().X), MVBuffer.View, FreqBuffer.View);
-            // BuildFreqKern(MVBuffer.Extent.ToIntIndex(), MVBuffer.View, FreqBuffer.View);
-            //}
-
-            BuildFreqKern(MVNormBuffer.Extent.ToIntIndex(), MVNormBuffer.View, FreqBuffer.View);
-
-            CalcEntropyKern(
-                FreqBuffer.Extent.ToIntIndex(),
-                FreqBuffer.View,
-                EntropyBuffer.View,
-                dataVector.GetLength(0)
-            );
-            double answer = EntropyBuffer.GetAsArray1D()[0] / Math.Log(LOG_BASE);
-            // print1d(MVNormBuffer.GetAsArray1D());
-            accelerate.Dispose();
-            return answer;
-        }
-
-        public double mergeArrays(double[] firstVector, double[] secondVector, ref double[] outputVector)
-        {
-            using Accelerator accelerate = this.dev.CreateAccelerator(this.context);
-            using var FirstBuffer = accelerate.Allocate1D<double>(new Index1D(firstVector.GetLength(0)));
-            using var SecondBuffer = accelerate.Allocate1D<double>(new Index1D(secondVector.GetLength(0)));
-
-            using var FirstNormBuffer = accelerate.Allocate1D<double>(new Index1D(firstVector.GetLength(0)));
-            using var SecondNormBuffer = accelerate.Allocate1D<double>(new Index1D(secondVector.GetLength(0)));
-
-            using var outputBuffer = accelerate.Allocate1D<double>(new Index1D(outputVector.GetLength(0)));
-
-            FirstBuffer.CopyFromCPU(firstVector);
-            SecondBuffer.CopyFromCPU(secondVector);
-
-            using var FirstMaxVal = accelerate.Allocate1D<int>(new Index1D(1));
-            using var FirstMinVal = accelerate.Allocate1D<int>(new Index1D(1));
-
-            using var SecondMaxVal = accelerate.Allocate1D<int>(new Index1D(1));
-            using var SecondMinVal = accelerate.Allocate1D<int>(new Index1D(1));
-
-            using var StateCount = accelerate.Allocate1D<int>(new Index1D(1));
-
-            var GetMaxValKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>
-            >(GetMaxMinValKernal);
-            var InitMaxMinKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>
-            >(InitMaxMinKernel);
-            var normalizeArrayKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>
-            >(normalizeArrayKernel);
-            var mergeArraysKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>,
-                int,
-                int
-            >(mergeArraysKernel);
-
-            InitMaxMinKern(FirstMaxVal.Extent.ToIntIndex(), FirstBuffer.View, FirstMaxVal.View, FirstMinVal.View);
-            GetMaxValKern(FirstBuffer.Extent.ToIntIndex(), FirstBuffer.View, FirstMaxVal.View, FirstMinVal.View);
-
-            InitMaxMinKern(SecondMaxVal.Extent.ToIntIndex(), SecondBuffer.View, SecondMaxVal.View, SecondMinVal.View);
-            GetMaxValKern(SecondBuffer.Extent.ToIntIndex(), SecondBuffer.View, SecondMaxVal.View, SecondMinVal.View);
-
-            normalizeArrayKern(FirstBuffer.Extent.ToIntIndex(), FirstBuffer.View, FirstNormBuffer.View, FirstMinVal);
-            normalizeArrayKern(
-                SecondBuffer.Extent.ToIntIndex(),
-                SecondBuffer.View,
-                SecondNormBuffer.View,
-                SecondMinVal
-            );
-            int firstnumstates = FirstMaxVal.GetAsArray1D()[0];
-            int secondnumstates = SecondMaxVal.GetAsArray1D()[0];
-
-            using var StateMap = accelerate.Allocate1D<int>(new Index1D((firstnumstates + 1) * (secondnumstates + 1)));
-            //outputVector = outputBuffer.GetAsArray1D();
-            mergeArraysKern(
-                StateCount.Extent.ToIntIndex(),
-                FirstNormBuffer.View,
-                SecondNormBuffer.View,
-                StateMap.View,
-                outputBuffer.View,
-                StateCount.View,
-                firstnumstates,
-                FirstNormBuffer.Extent.ToIntIndex().X
-            );
-            outputVector = outputBuffer.GetAsArray1D();
-            // print1d(outputVector);
-            double answer = StateCount.GetAsArray1D()[0];
-            accelerate.Dispose();
-            return answer;
-        }
-
-        public double calculateJointEntropy(double[] firstVector, double[] secondVector)
-        {
-            using Accelerator accelerate = this.dev.CreateAccelerator(this.context);
-            using var FirstBuffer = accelerate.Allocate1D<double>(new Index1D(firstVector.GetLength(0)));
-            using var SecondBuffer = accelerate.Allocate1D<double>(new Index1D(secondVector.GetLength(0)));
-
-            FirstBuffer.CopyFromCPU(firstVector);
-            SecondBuffer.CopyFromCPU(secondVector);
-
-            using var FirstNormBuffer = accelerate.Allocate1D<double>(new Index1D(firstVector.GetLength(0)));
-            using var SecondNormBuffer = accelerate.Allocate1D<double>(new Index1D(secondVector.GetLength(0)));
-            using var FirstMaxVal = accelerate.Allocate1D<int>(new Index1D(1));
-            using var FirstMinVal = accelerate.Allocate1D<int>(new Index1D(1));
-
-            using var SecondMaxVal = accelerate.Allocate1D<int>(new Index1D(1));
-            using var SecondMinVal = accelerate.Allocate1D<int>(new Index1D(1));
-            using var EntropyBuffer = accelerate.Allocate1D<double>(new Index1D(1));
-            using var TestBuffer = accelerate.Allocate1D<double>(new Index1D(1));
-
-            var GetMaxValKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>
-            >(GetMaxMinValKernal);
-            var InitMaxMinKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>
-            >(InitMaxMinKernel);
-            var normalizeArrayKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>
-            >(normalizeArrayKernel);
-
-            var BuildJointFreqKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView2D<double, Stride2D.DenseX>
-            >(BuildJointFreqKernel);
-
-            var setBuffToValue2DKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index2D,
-                ArrayView2D<double, Stride2D.DenseX>,
-                double
-            >(setBuffToValue2DKernal);
-
-            var CalcJointEntropyKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index2D,
-                ArrayView2D<double, Stride2D.DenseX>,
-                ArrayView1D<double, Stride1D.Dense>,
-                int
-            >(CalcJointEntropyKernel);
-
-            var IndexedCalcJointEntropyKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView2D<double, Stride2D.DenseX>,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>,
-                int
-            >(IndexedCalcJointEntropyKernel);
-
-            var setBuffToValueDoubleKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                double
-            >(setBuffToValueDoubleKernal);
-
-            var BuildFreqKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>
-            >(BuildFreqAdjustedKernel);
-            InitMaxMinKern(FirstMaxVal.Extent.ToIntIndex(), FirstBuffer.View, FirstMaxVal.View, FirstMinVal.View);
-            GetMaxValKern(FirstBuffer.Extent.ToIntIndex(), FirstBuffer.View, FirstMaxVal.View, FirstMinVal.View);
-            //Console.WriteLine(FirstMaxVal.GetAsArray1D()[0]);
-            InitMaxMinKern(SecondMaxVal.Extent.ToIntIndex(), SecondBuffer.View, SecondMaxVal.View, SecondMinVal.View);
-            GetMaxValKern(SecondBuffer.Extent.ToIntIndex(), SecondBuffer.View, SecondMaxVal.View, SecondMinVal.View);
-            normalizeArrayKern(FirstBuffer.Extent.ToIntIndex(), FirstBuffer.View, FirstNormBuffer.View, FirstMinVal);
-            normalizeArrayKern(
-                SecondBuffer.Extent.ToIntIndex(),
-                SecondBuffer.View,
-                SecondNormBuffer.View,
-                SecondMinVal
-            );
-            // Console.WriteLine("Norms");
-            // print1d(FirstNormBuffer.GetAsArray1D());
-            // print1d(SecondNormBuffer.GetAsArray1D());
-
-            // csharpier-ignore-start
-            // Reusing these variables from before because we don't need the original values anymore now that everything is normalized
-            InitMaxMinKern(FirstMaxVal.Extent.ToIntIndex(), FirstNormBuffer.View, FirstMaxVal.View, FirstMinVal.View);
-            GetMaxValKern(FirstNormBuffer.Extent.ToIntIndex(), FirstNormBuffer.View, FirstMaxVal.View, FirstMinVal.View);
-            InitMaxMinKern(SecondMaxVal.Extent.ToIntIndex(), SecondNormBuffer.View, SecondMaxVal.View, SecondMinVal.View);
-            GetMaxValKern(SecondNormBuffer.Extent.ToIntIndex(), SecondNormBuffer.View, SecondMaxVal.View, SecondMinVal.View);
-            // csharpier-ignore-end
-
-            int firstMaxVal = FirstMaxVal.GetAsArray1D()[0];
-            int secondMaxVal = SecondMaxVal.GetAsArray1D()[0];
-
-            using var JointBuffer = accelerate.Allocate2DDenseX<double>(new Index2D(firstMaxVal + 1, secondMaxVal + 1));
-            // Console.WriteLine($"JointBuffer Size: {JointBuffer.Length}");
-
-            // setBuffToValue2DKern(JointBuffer.Extent.ToIntIndex(), JointBuffer.View, 0.0);
-            // setBuffToValueDoubleKern(FirstBuffer.Extent.ToIntIndex(), FirstBuffer.View, 0.0);
-            // setBuffToValueDoubleKern(FirstBuffer.Extent.ToIntIndex(), FirstBuffer.View, 0.0);
-
-            BuildJointFreqKern(
-                SecondNormBuffer.Extent.ToIntIndex(),
-                FirstNormBuffer.View,
-                SecondNormBuffer.View,
-                JointBuffer.View
-            );
-            // print1d(FirstNormBuffer.GetAsArray1D());
-            // print1d(SecondNormBuffer.GetAsArray1D());
-            // print2d(JointBuffer.GetAsArray2D());
-
-            //CalcJointEntropyKern(JointBuffer.Extent.ToIntIndex(), JointBuffer.View, EntropyBuffer.View, firstVector.GetLength(0));
-            IndexedCalcJointEntropyKern(
-                SecondNormBuffer.Extent.ToIntIndex(),
-                JointBuffer.View,
-                FirstNormBuffer.View,
-                SecondNormBuffer.View,
-                EntropyBuffer.View,
-                firstVector.GetLength(0)
-            );
-
-            // Console.WriteLine("joint");
-            // Console.WriteLine(EntropyBuffer.GetAsArray1D()[0]);
-            // Console.WriteLine(TestBuffer.GetAsArray1D()[0]);
-
-            double answer = EntropyBuffer.GetAsArray1D()[0] / Math.Log(LOG_BASE);
-            accelerate.Dispose();
-            return answer;
-        }
-
-        /// <summary>
-        /// Alternate implementation of JointEntropy using a 1D Buffer (szudzik for hashing) instead of a 2D Buffer.
-        /// </summary>
-        /// <param name="firstVector"></param>
-        /// <param name="secondVector"></param>
-        /// <returns></returns>
-        public double calculateJointEntropy2(double[] firstVector, double[] secondVector)
-        {
-            using Accelerator accelerate = this.dev.CreateAccelerator(this.context);
-            using var FirstBuffer = accelerate.Allocate1D<double>(new Index1D(firstVector.GetLength(0)));
-            using var SecondBuffer = accelerate.Allocate1D<double>(new Index1D(secondVector.GetLength(0)));
-
-            FirstBuffer.CopyFromCPU(firstVector);
-            SecondBuffer.CopyFromCPU(secondVector);
-
-            using var FirstNormBuffer = accelerate.Allocate1D<double>(new Index1D(firstVector.GetLength(0)));
-            using var SecondNormBuffer = accelerate.Allocate1D<double>(new Index1D(secondVector.GetLength(0)));
-            using var FirstMaxVal = accelerate.Allocate1D<int>(new Index1D(1));
-            using var FirstMinVal = accelerate.Allocate1D<int>(new Index1D(1));
-
-            using var SecondMaxVal = accelerate.Allocate1D<int>(new Index1D(1));
-            using var SecondMinVal = accelerate.Allocate1D<int>(new Index1D(1));
-            using var EntropyBuffer = accelerate.Allocate1D<double>(new Index1D(1));
-            using var TestBuffer = accelerate.Allocate1D<double>(new Index1D(1));
-
-            var GetMaxValKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>
-            >(GetMaxMinValKernal);
-
-            var InitMaxMinKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>
-            >(InitMaxMinKernel);
-
-            var normalizeArrayKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>
-            >(normalizeArrayKernel);
-
-            var BuildJointFreqKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView<double>,
-                ArrayView<double>,
-                ArrayView<double>,
-                long
-            >(BuildJointFreqKernel2);
-
-            var setBuffToValue2DKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index2D,
-                ArrayView2D<double, Stride2D.DenseX>,
-                double
-            >(setBuffToValue2DKernal);
-
-            var CalcJointEntropyKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index2D,
-                ArrayView2D<double, Stride2D.DenseX>,
-                ArrayView1D<double, Stride1D.Dense>,
-                int
-            >(CalcJointEntropyKernel);
-
-            var IndexedCalcJointEntropyKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView<double>,
-                ArrayView<double>,
-                ArrayView<double>,
-                ArrayView<double>,
-                int,
-                long
-            >(IndexedCalcJointEntropyKernel2);
-
-            var setBuffToValueDoubleKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                double
-            >(setBuffToValueDoubleKernal);
-
-            var BuildFreqKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>
-            >(BuildFreqAdjustedKernel);
-
-            InitMaxMinKern(FirstMaxVal.Extent.ToIntIndex(), FirstBuffer.View, FirstMaxVal.View, FirstMinVal.View);
-            GetMaxValKern(FirstBuffer.Extent.ToIntIndex(), FirstBuffer.View, FirstMaxVal.View, FirstMinVal.View);
-            InitMaxMinKern(SecondMaxVal.Extent.ToIntIndex(), SecondBuffer.View, SecondMaxVal.View, SecondMinVal.View);
-            GetMaxValKern(SecondBuffer.Extent.ToIntIndex(), SecondBuffer.View, SecondMaxVal.View, SecondMinVal.View);
-
-            normalizeArrayKern(FirstBuffer.Extent.ToIntIndex(), FirstBuffer.View, FirstNormBuffer.View, FirstMinVal);
-            normalizeArrayKern(
-                SecondBuffer.Extent.ToIntIndex(),
-                SecondBuffer.View,
-                SecondNormBuffer.View,
-                SecondMinVal
-            );
-            // Console.WriteLine($"FirstNormBuffer: [{string.Join(", ", FirstNormBuffer.GetAsArray1D())}]");
-            // Console.WriteLine($"SecondNormBuffer: [{string.Join(", ", SecondNormBuffer.GetAsArray1D())}]");
-
-            // csharpier-ignore-start
-            // Reusing these variables from before because we don't need the original values anymore now that everything is normalized
-            InitMaxMinKern(FirstMaxVal.Extent.ToIntIndex(), FirstNormBuffer.View, FirstMaxVal.View, FirstMinVal.View);
-            GetMaxValKern(FirstNormBuffer.Extent.ToIntIndex(), FirstNormBuffer.View, FirstMaxVal.View, FirstMinVal.View);
-            InitMaxMinKern(SecondMaxVal.Extent.ToIntIndex(), SecondNormBuffer.View, SecondMaxVal.View, SecondMinVal.View);
-            GetMaxValKern(SecondNormBuffer.Extent.ToIntIndex(), SecondNormBuffer.View, SecondMaxVal.View, SecondMinVal.View);
-            // csharpier-ignore-end
-
-            int firstMaxVal = FirstMaxVal.GetAsArray1D()[0];
-            int secondMaxVal = SecondMaxVal.GetAsArray1D()[0];
-            // Console.WriteLine($"FirstMaxVal = {firstMaxVal:n0}");
-            // Console.WriteLine($"SecondMaxVal = {secondMaxVal:n0}");
-            int firstMinVal = FirstMinVal.GetAsArray1D()[0];
-            int secondMinVal = SecondMinVal.GetAsArray1D()[0];
-            // Console.WriteLine($"FirstMinVal = {firstMinVal:n0}");
-            // Console.WriteLine($"SecondMinVal = {secondMinVal:n0}");
-
-            long maxIndexVal = szudzikPair(firstMaxVal, secondMaxVal);
-            byte maxOrderMagnitude = (byte)maxIndexVal.ToString().Length;
-            // Console.WriteLine($"MaxPossiblePair = {maxIndexVal:n0} with {maxOrderMagnitude} digits");
-            long minIndexVal = szudzikPair(firstMinVal, secondMinVal);
-            // Console.WriteLine($"MinPossiblePair = {minIndexVal:n0}");
-            using var JointBuffer = accelerate.Allocate1D<double>(maxIndexVal - minIndexVal);
-            // JointBuffer.MemSetToZero();
-            // Console.WriteLine($"JointBuffer Size: {JointBuffer.Length:n0}");
-
-            BuildJointFreqKern(
-                SecondNormBuffer.Extent.ToIntIndex(),
-                FirstNormBuffer.View,
-                SecondNormBuffer.View,
-                JointBuffer.View,
-                minIndexVal
-            );
-
-            IndexedCalcJointEntropyKern(
-                SecondNormBuffer.Extent.ToIntIndex(),
-                JointBuffer.View,
-                FirstNormBuffer.View,
-                SecondNormBuffer.View,
-                EntropyBuffer.View,
-                firstVector.GetLength(0),
-                minIndexVal
-            );
-
-            double answer = EntropyBuffer.GetAsArray1D()[0] / Math.Log(LOG_BASE);
-            accelerate.Dispose();
-            return answer;
-        }
-
-        // TODO: Function not finished, just testing for now
-        public double calculateJointEntropy3(double[] firstVector, double[] secondVector)
-        {
-            using Accelerator accelerate = this.dev.CreateAccelerator(this.context);
-            using var FirstBuffer = accelerate.Allocate1D<double>(new Index1D(firstVector.GetLength(0)));
-            using var SecondBuffer = accelerate.Allocate1D<double>(new Index1D(secondVector.GetLength(0)));
-
-            FirstBuffer.CopyFromCPU(firstVector);
-            SecondBuffer.CopyFromCPU(secondVector);
-
-            using var FirstNormBuffer = accelerate.Allocate1D<double>(new Index1D(firstVector.GetLength(0)));
-            using var SecondNormBuffer = accelerate.Allocate1D<double>(new Index1D(secondVector.GetLength(0)));
-            using var FirstMaxVal = accelerate.Allocate1D<int>(new Index1D(1));
-            using var FirstMinVal = accelerate.Allocate1D<int>(new Index1D(1));
-
-            using var SecondMaxVal = accelerate.Allocate1D<int>(new Index1D(1));
-            using var SecondMinVal = accelerate.Allocate1D<int>(new Index1D(1));
-            using var EntropyBuffer = accelerate.Allocate1D<double>(new Index1D(1));
-            using var TestBuffer = accelerate.Allocate1D<double>(new Index1D(1));
-
-            var GetMaxValKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>
-            >(GetMaxMinValKernal);
-            var InitMaxMinKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>
-            >(InitMaxMinKernel);
-            var normalizeArrayKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>
-            >(normalizeArrayKernel);
-
-            var BuildJointFreqKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView2D<double, Stride2D.DenseX>
-            >(BuildJointFreqKernel3);
-
-            var setBuffToValue2DKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index2D,
-                ArrayView2D<double, Stride2D.DenseX>,
-                double
-            >(setBuffToValue2DKernal);
-
-            var CalcJointEntropyKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index2D,
-                ArrayView2D<double, Stride2D.DenseX>,
-                ArrayView1D<double, Stride1D.Dense>,
-                int
-            >(CalcJointEntropyKernel);
-
-            var IndexedCalcJointEntropyKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView2D<double, Stride2D.DenseX>,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>,
-                int
-            >(IndexedCalcJointEntropyKernel3);
-
-            var setBuffToValueDoubleKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                double
-            >(setBuffToValueDoubleKernal);
-
-            var BuildFreqKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>
-            >(BuildFreqAdjustedKernel);
-
-            InitMaxMinKern(FirstMaxVal.Extent.ToIntIndex(), FirstBuffer.View, FirstMaxVal.View, FirstMinVal.View);
-            GetMaxValKern(FirstBuffer.Extent.ToIntIndex(), FirstBuffer.View, FirstMaxVal.View, FirstMinVal.View);
-            InitMaxMinKern(SecondMaxVal.Extent.ToIntIndex(), SecondBuffer.View, SecondMaxVal.View, SecondMinVal.View);
-            GetMaxValKern(SecondBuffer.Extent.ToIntIndex(), SecondBuffer.View, SecondMaxVal.View, SecondMinVal.View);
-            // csharpier-ignore-start
-            normalizeArrayKern(FirstBuffer.Extent.ToIntIndex(), FirstBuffer.View, FirstNormBuffer.View, FirstMinVal);
-            normalizeArrayKern(SecondBuffer.Extent.ToIntIndex(), SecondBuffer.View, SecondNormBuffer.View, SecondMinVal);
-            // csharpier-ignore-end
-
-            int firstMaxVal = FirstMaxVal.GetAsArray1D()[0];
-            int secondMaxVal = SecondMaxVal.GetAsArray1D()[0];
-            Console.WriteLine($"FirstMaxVal = {firstMaxVal:n0}");
-            Console.WriteLine($"SecondMaxVal = {secondMaxVal:n0}");
-
-            long maxIndexVal = szudzikPair(firstMaxVal, secondMaxVal);
-            byte maxOrderMagnitude = (byte)maxIndexVal.ToString().Length;
-            long maxIndexValSqrt = (long)Math.Ceiling(Math.Sqrt(maxIndexVal));
-            Console.WriteLine(
-                $"MaxPossiblePair = {maxIndexVal:n0} with {maxOrderMagnitude} digits, with matrix rank {maxIndexValSqrt:n0}"
-            );
-            using var JointBuffer = accelerate.Allocate2DDenseX<double>(
-                new LongIndex2D(maxIndexValSqrt, maxIndexValSqrt)
-            );
-
-            BuildJointFreqKern(
-                SecondNormBuffer.Extent.ToIntIndex(),
-                FirstNormBuffer.View,
-                SecondNormBuffer.View,
-                JointBuffer.View
-            );
-
-            IndexedCalcJointEntropyKern(
-                SecondNormBuffer.Extent.ToIntIndex(),
-                JointBuffer.View,
-                FirstNormBuffer.View,
-                SecondNormBuffer.View,
-                EntropyBuffer.View,
-                firstVector.GetLength(0)
-            );
-
-            double answer = EntropyBuffer.GetAsArray1D()[0] / Math.Log(LOG_BASE);
-            accelerate.Dispose();
-            return answer;
-        }
-
-        public double calculateConditionalEntropy(double[] firstVector, double[] secondVector)
-        {
-            //// NOTES: Need to account for the NaN values, have them add a 1/n addition to the entropy
-            using Accelerator accelerate = this.dev.CreateAccelerator(this.context);
-            using var FirstBuffer = accelerate.Allocate1D<double>(new Index1D(firstVector.GetLength(0)));
-            using var SecondBuffer = accelerate.Allocate1D<double>(new Index1D(secondVector.GetLength(0)));
-            double answer;
-            FirstBuffer.CopyFromCPU(firstVector);
-            SecondBuffer.CopyFromCPU(secondVector);
-
-            using var FirstNormBuffer = accelerate.Allocate1D<double>(new Index1D(firstVector.GetLength(0)));
-            using var SecondNormBuffer = accelerate.Allocate1D<double>(new Index1D(secondVector.GetLength(0)));
-            using var FirstMaxVal = accelerate.Allocate1D<int>(new Index1D(1));
-            using var FirstMinVal = accelerate.Allocate1D<int>(new Index1D(1));
-
-            using var SecondMaxVal = accelerate.Allocate1D<int>(new Index1D(1));
-            using var SecondMinVal = accelerate.Allocate1D<int>(new Index1D(1));
-            using var EntropyBuffer = accelerate.Allocate1D<double>(new Index1D(1));
-            using var TestBuffer = accelerate.Allocate1D<double>(new Index1D(1));
-
-            var GetMaxValKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>
-            >(TestGetMaxMinValKernal);
-            var InitMaxMinKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>
-            >(InitMaxMinKernel);
-            var normalizeArrayKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>
-            >(normalizeArrayKernel);
-
-            var BuildJointFreqKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView2D<double, Stride2D.DenseX>
-            >(BuildJointFreqKernel);
-
-            var setBuffToValue2DKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index2D,
-                ArrayView2D<double, Stride2D.DenseX>,
-                double
-            >(setBuffToValue2DKernal);
-
-            var CalcConditionalEntropyKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index2D,
-                ArrayView2D<double, Stride2D.DenseX>,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>,
-                int
-            >(CalcConditionalEntropyKernel);
-
-            var IndexedCalcConditionalEntropyKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView2D<double, Stride2D.DenseX>,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>,
-                int
-            >(IndexedCalcConditionalEntropyKernel);
-
-            var setBuffToValueDoubleKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                double
-            >(setBuffToValueDoubleKernal);
-
-            var BuildFreqKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>
-            >(BuildFreqAdjKernel);
-
-            InitMaxMinKern(FirstMaxVal.Extent.ToIntIndex(), FirstBuffer.View, FirstMaxVal.View, FirstMinVal.View);
-            GetMaxValKern(FirstBuffer.Extent.ToIntIndex(), FirstBuffer.View, FirstMaxVal.View, FirstMinVal.View);
-            //Console.WriteLine(FirstMaxVal.GetAsArray1D()[0]);
-            InitMaxMinKern(SecondMaxVal.Extent.ToIntIndex(), SecondBuffer.View, SecondMaxVal.View, SecondMinVal.View);
-            //print1d(SecondBuffer.GetAsArray1D());
-            GetMaxValKern(SecondBuffer.Extent.ToIntIndex(), SecondBuffer.View, SecondMaxVal.View, SecondMinVal.View);
-            //print1d(SecondMinVal.GetAsArray1D());
-
-            normalizeArrayKern(FirstBuffer.Extent.ToIntIndex(), FirstBuffer.View, FirstNormBuffer.View, FirstMinVal);
-            normalizeArrayKern(
-                SecondBuffer.Extent.ToIntIndex(),
-                SecondBuffer.View,
-                SecondNormBuffer.View,
-                SecondMinVal
-            );
-            // csharpier-ignore-start
-            // Reusing these variables from before because we don't need the original values anymore now that everything is normalized
-            InitMaxMinKern(FirstMaxVal.Extent.ToIntIndex(), FirstNormBuffer.View, FirstMaxVal.View, FirstMinVal.View);
-            GetMaxValKern(FirstNormBuffer.Extent.ToIntIndex(), FirstNormBuffer.View, FirstMaxVal.View, FirstMinVal.View);
-            InitMaxMinKern(SecondMaxVal.Extent.ToIntIndex(), SecondNormBuffer.View, SecondMaxVal.View, SecondMinVal.View);
-            GetMaxValKern(SecondNormBuffer.Extent.ToIntIndex(), SecondNormBuffer.View, SecondMaxVal.View, SecondMinVal.View);
-            // csharpier-ignore-end
-
-            // Console.WriteLine("Norms");
-            // print1d(FirstNormBuffer.GetAsArray1D());
-            // print1d(SecondNormBuffer.GetAsArray1D());
-            int firstnumstates = FirstMaxVal.GetAsArray1D()[0];
-            int secondnumstates = SecondMaxVal.GetAsArray1D()[0];
-            // Console.WriteLine(firstnumstates);
-            // Console.WriteLine(secondnumstates);
-
-            using var JointBuffer = accelerate.Allocate2DDenseX<double>(
-                new Index2D(firstnumstates + 1, secondnumstates + 1)
-            );
-            answer = EntropyBuffer.GetAsArray1D()[0] / Math.Log(LOG_BASE);
-
-            using var FirstCountMap = accelerate.Allocate1D<double>(new Index1D(firstnumstates));
-            using var SecondCountMap = accelerate.Allocate1D<double>(new Index1D(secondnumstates));
-            //setBuffToValue2DKern(JointBuffer.Extent.ToIntIndex(), JointBuffer.View, 0.0);
-            // setBuffToValueDoubleKern(FirstBuffer.Extent.ToIntIndex(), FirstBuffer.View, 0.0);
-            // setBuffToValueDoubleKern(FirstBuffer.Extent.ToIntIndex(), FirstBuffer.View, 0.0);
-            //Console.WriteLine(EntropyBuffer.GetAsArray1D()[0]);
-            answer = EntropyBuffer.GetAsArray1D()[0] / Math.Log(LOG_BASE);
-            // Console.WriteLine("First and second norm:");
-            //print1d(SecondMinVal.GetAsArray1D());
-            //print1d(SecondNormBuffer.GetAsArray1D());
-            //Console.ReadLine();
-
-            BuildJointFreqKern(
-                SecondNormBuffer.Extent.ToIntIndex(),
-                FirstNormBuffer.View,
-                SecondNormBuffer.View,
-                JointBuffer.View
-            );
-            // print2d(JointBuffer.GetAsArray2D());
-            //Console.WriteLine(EntropyBuffer.GetAsArray1D()[0]);
-            answer = EntropyBuffer.GetAsArray1D()[0] / Math.Log(LOG_BASE);
-            BuildFreqKern(SecondNormBuffer.Extent.ToIntIndex(), SecondNormBuffer.View, SecondCountMap.View);
-            //Console.WriteLine(EntropyBuffer.GetAsArray1D()[0]);
-            // Console.WriteLine("JOINT");
-            //print2d(JointBuffer.GetAsArray2D());
-            // Console.WriteLine("SecondCountMap");
-            //print1d(SecondCountMap.GetAsArray1D());
-            // /CalcConditionalEntropyKern(JointBuffer.Extent.ToIntIndex(), JointBuffer.View, SecondCountMap.View, EntropyBuffer.View, firstVector.GetLength(0));
-            answer = EntropyBuffer.GetAsArray1D()[0] / Math.Log(LOG_BASE);
-            IndexedCalcConditionalEntropyKern(
-                SecondNormBuffer.Extent.ToIntIndex(),
-                JointBuffer.View,
-                SecondCountMap.View,
-                FirstNormBuffer.View,
-                SecondNormBuffer.View,
-                EntropyBuffer.View,
-                firstVector.GetLength(0)
-            );
-
-            // Console.WriteLine("TESTTT");
-            // Console.WriteLine(EntropyBuffer.GetAsArray1D()[0]);
-            // Console.WriteLine(TestBuffer.GetAsArray1D()[0]);
-
-            // Console.WriteLine("ENTROPY ARR");
-            // print1d(EntropyBuffer.GetAsArray1D());
-            //double answer = EntropyBuffer.GetAsArray1D()[0] / Math.Log(LOG_BASE);
-            answer = EntropyBuffer.GetAsArray1D()[0] / Math.Log(LOG_BASE);
-            // print1d(FirstNormBuffer.GetAsArray1D());
-            // print1d(SecondNormBuffer.GetAsArray1D());
-            accelerate.Dispose();
-            return answer;
-        }
-
-        public double calculateConditionalEntropyAdjusted(double[] firstVector, double[] secondVector, int nanvals)
-        {
-            //// NOTES: Need to account for the NaN values, have them add a 1/n addition to the entropy
-            using Accelerator accelerate = this.dev.CreateAccelerator(this.context);
-            using var FirstBuffer = accelerate.Allocate1D<double>(new Index1D(firstVector.GetLength(0)));
-            using var SecondBuffer = accelerate.Allocate1D<double>(new Index1D(secondVector.GetLength(0)));
-            double answer;
-            FirstBuffer.CopyFromCPU(firstVector);
-            SecondBuffer.CopyFromCPU(secondVector);
-
-            using var FirstNormBuffer = accelerate.Allocate1D<double>(new Index1D(firstVector.GetLength(0)));
-            using var SecondNormBuffer = accelerate.Allocate1D<double>(new Index1D(secondVector.GetLength(0)));
-            using var FirstMaxVal = accelerate.Allocate1D<int>(new Index1D(1));
-            using var FirstMinVal = accelerate.Allocate1D<int>(new Index1D(1));
-
-            using var SecondMaxVal = accelerate.Allocate1D<int>(new Index1D(1));
-            using var SecondMinVal = accelerate.Allocate1D<int>(new Index1D(1));
-            using var EntropyBuffer = accelerate.Allocate1D<double>(new Index1D(1));
-            using var TestBuffer = accelerate.Allocate1D<double>(new Index1D(1));
-
-            var GetMaxValKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>
-            >(TestGetMaxMinValKernal);
-            var InitMaxMinKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>
-            >(InitMaxMinKernel);
-            var normalizeArrayKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>
-            >(normalizeArrayKernel);
-
-            var BuildJointFreqKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView2D<double, Stride2D.DenseX>
-            >(BuildJointFreqKernel);
-
-            var setBuffToValue2DKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index2D,
-                ArrayView2D<double, Stride2D.DenseX>,
-                double
-            >(setBuffToValue2DKernal);
-
-            var CalcConditionalEntropyKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index2D,
-                ArrayView2D<double, Stride2D.DenseX>,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>,
-                int
-            >(CalcConditionalEntropyKernel);
-
-            var IndexedCalcConditionalEntropyKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView2D<double, Stride2D.DenseX>,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>,
-                int
-            >(IndexedCalcConditionalEntropyKernel);
-
-            var setBuffToValueDoubleKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                double
-            >(setBuffToValueDoubleKernal);
-
-            var BuildFreqKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>
-            >(BuildFreqAdjKernel);
-            // Console.WriteLine("NAN FLOOR");
-            // Console.WriteLine(Math.Floor(Double.NaN));
-            InitMaxMinKern(FirstMaxVal.Extent.ToIntIndex(), FirstBuffer.View, FirstMaxVal.View, FirstMinVal.View);
-            GetMaxValKern(FirstBuffer.Extent.ToIntIndex(), FirstBuffer.View, FirstMaxVal.View, FirstMinVal.View);
-            //Console.WriteLine(FirstMaxVal.GetAsArray1D()[0]);
-            InitMaxMinKern(SecondMaxVal.Extent.ToIntIndex(), SecondBuffer.View, SecondMaxVal.View, SecondMinVal.View);
-            //print1d(SecondMinVal.GetAsArray1D());
-            GetMaxValKern(SecondBuffer.Extent.ToIntIndex(), SecondBuffer.View, SecondMaxVal.View, SecondMinVal.View);
-            //print1d(SecondMinVal.GetAsArray1D());
-
-            normalizeArrayKern(FirstBuffer.Extent.ToIntIndex(), FirstBuffer.View, FirstNormBuffer.View, FirstMinVal);
-            normalizeArrayKern(
-                SecondBuffer.Extent.ToIntIndex(),
-                SecondBuffer.View,
-                SecondNormBuffer.View,
-                SecondMinVal
-            );
-            // Console.WriteLine("Norms");
-            // print1d(FirstNormBuffer.GetAsArray1D());
-            // print1d(SecondNormBuffer.GetAsArray1D());
-            int firstnumstates = FirstMaxVal.GetAsArray1D()[0];
-            int secondnumstates = SecondMaxVal.GetAsArray1D()[0];
-            // Console.WriteLine(firstnumstates);
-            // Console.WriteLine(secondnumstates);
-
-            using var JointBuffer = accelerate.Allocate2DDenseX<double>(
-                new Index2D(firstnumstates + 1, secondnumstates + 1)
-            );
-            //answer = EntropyBuffer.GetAsArray1D()[0] / Math.Log(LOG_BASE);
-
-            using var FirstCountMap = accelerate.Allocate1D<double>(new Index1D(firstnumstates));
-            using var SecondCountMap = accelerate.Allocate1D<double>(new Index1D(secondnumstates));
-            //setBuffToValue2DKern(JointBuffer.Extent.ToIntIndex(), JointBuffer.View, 0.0);
-            // setBuffToValueDoubleKern(FirstBuffer.Extent.ToIntIndex(), FirstBuffer.View, 0.0);
-            // setBuffToValueDoubleKern(FirstBuffer.Extent.ToIntIndex(), FirstBuffer.View, 0.0);
-            //Console.WriteLine(EntropyBuffer.GetAsArray1D()[0]);
-            //answer = EntropyBuffer.GetAsArray1D()[0] / Math.Log(LOG_BASE);
-            // Console.WriteLine("First and second norm:");
-            // print1d(SecondMinVal.GetAsArray1D());
-            // print1d(SecondNormBuffer.GetAsArray1D());
-            // Console.ReadLine();
-
-            BuildJointFreqKern(
-                SecondNormBuffer.Extent.ToIntIndex(),
-                FirstNormBuffer.View,
-                SecondNormBuffer.View,
-                JointBuffer.View
-            );
-            //print2d(JointBuffer.GetAsArray2D());
-            //Console.WriteLine(EntropyBuffer.GetAsArray1D()[0]);
-            //answer = EntropyBuffer.GetAsArray1D()[0] / Math.Log(LOG_BASE);
-            BuildFreqKern(SecondNormBuffer.Extent.ToIntIndex(), SecondNormBuffer.View, SecondCountMap.View);
-            //Console.WriteLine(EntropyBuffer.GetAsArray1D()[0]);
-            // Console.WriteLine("JOINT");
-            //print2d(JointBuffer.GetAsArray2D());
-            // Console.WriteLine("SecondCountMap");
-            //print1d(SecondCountMap.GetAsArray1D());
-            // /CalcConditionalEntropyKern(JointBuffer.Extent.ToIntIndex(), JointBuffer.View, SecondCountMap.View, EntropyBuffer.View, firstVector.GetLength(0));
-            //answer = EntropyBuffer.GetAsArray1D()[0] / Math.Log(LOG_BASE);
-            IndexedCalcConditionalEntropyKern(
-                SecondNormBuffer.Extent.ToIntIndex(),
-                JointBuffer.View,
-                SecondCountMap.View,
-                FirstNormBuffer.View,
-                SecondNormBuffer.View,
-                EntropyBuffer.View,
-                firstVector.GetLength(0)
-            );
-
-            // Console.WriteLine("TESTTT");
-            // Console.WriteLine(EntropyBuffer.GetAsArray1D()[0]);
-            // Console.WriteLine(TestBuffer.GetAsArray1D()[0]);
-
-            // Console.WriteLine("ENTROPY ARR");
-            // print1d(EntropyBuffer.GetAsArray1D());
-            //double answer = EntropyBuffer.GetAsArray1D()[0] / Math.Log(LOG_BASE);
-            answer = EntropyBuffer.GetAsArray1D()[0] / Math.Log(LOG_BASE);
-            accelerate.Dispose();
-            return answer;
-        }
-
-        public double calculateMutualInformation(double[] firstVector, double[] secondVector)
-        {
-            using Accelerator accelerate = this.dev.CreateAccelerator(this.context);
-            using var FirstBuffer = accelerate.Allocate1D<double>(new Index1D(firstVector.GetLength(0)));
-            using var SecondBuffer = accelerate.Allocate1D<double>(new Index1D(secondVector.GetLength(0)));
-
-            FirstBuffer.CopyFromCPU(firstVector);
-            SecondBuffer.CopyFromCPU(secondVector);
-
-            using var FirstNormBuffer = accelerate.Allocate1D<double>(new Index1D(firstVector.GetLength(0)));
-            using var SecondNormBuffer = accelerate.Allocate1D<double>(new Index1D(secondVector.GetLength(0)));
-            using var FirstMaxVal = accelerate.Allocate1D<int>(new Index1D(1));
-            using var FirstMinVal = accelerate.Allocate1D<int>(new Index1D(1));
-
-            using var SecondMaxVal = accelerate.Allocate1D<int>(new Index1D(1));
-            using var SecondMinVal = accelerate.Allocate1D<int>(new Index1D(1));
-            using var EntropyBuffer = accelerate.Allocate1D<double>(new Index1D(1));
-
-            var GetMaxValKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>
-            >(GetMaxMinValKernal);
-            var InitMaxMinKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>
-            >(InitMaxMinKernel);
-            var normalizeArrayKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>
-            >(normalizeArrayKernel);
-
-            var BuildJointFreqKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView2D<double, Stride2D.DenseX>
-            >(BuildJointFreqKernel);
-
-            var setBuffToValue2DKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index2D,
-                ArrayView2D<double, Stride2D.DenseX>,
-                double
-            >(setBuffToValue2DKernal);
-
-            var CalcMIKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index2D,
-                ArrayView2D<double, Stride2D.DenseX>,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>,
-                int
-            >(CalcMIKernel);
-
-            var IndexedCalcMIKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView2D<double, Stride2D.DenseX>,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>,
-                int
-            >(IndexedCalcMIKernel);
-
-            var setBuffToValueDoubleKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                double
-            >(setBuffToValueDoubleKernal);
-
-            var BuildFreqKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>
-            >(BuildFreqAdjustedKernel);
-
-            InitMaxMinKern(FirstMaxVal.Extent.ToIntIndex(), FirstBuffer.View, FirstMaxVal.View, FirstMinVal.View);
-            GetMaxValKern(FirstBuffer.Extent.ToIntIndex(), FirstBuffer.View, FirstMaxVal.View, FirstMinVal.View);
-            //Console.WriteLine(FirstMaxVal.GetAsArray1D()[0]);
-            InitMaxMinKern(SecondMaxVal.Extent.ToIntIndex(), SecondBuffer.View, SecondMaxVal.View, SecondMinVal.View);
-            GetMaxValKern(SecondBuffer.Extent.ToIntIndex(), SecondBuffer.View, SecondMaxVal.View, SecondMinVal.View);
-
-            normalizeArrayKern(FirstBuffer.Extent.ToIntIndex(), FirstBuffer.View, FirstNormBuffer.View, FirstMinVal);
-            normalizeArrayKern(
-                SecondBuffer.Extent.ToIntIndex(),
-                SecondBuffer.View,
-                SecondNormBuffer.View,
-                SecondMinVal
-            );
-            // csharpier-ignore-start
-            // Reusing these variables from before because we don't need the original values anymore now that everything is normalized
-            InitMaxMinKern(FirstMaxVal.Extent.ToIntIndex(), FirstNormBuffer.View, FirstMaxVal.View, FirstMinVal.View);
-            GetMaxValKern(FirstNormBuffer.Extent.ToIntIndex(), FirstNormBuffer.View, FirstMaxVal.View, FirstMinVal.View);
-            InitMaxMinKern(SecondMaxVal.Extent.ToIntIndex(), SecondNormBuffer.View, SecondMaxVal.View, SecondMinVal.View);
-            GetMaxValKern(SecondNormBuffer.Extent.ToIntIndex(), SecondNormBuffer.View, SecondMaxVal.View, SecondMinVal.View);
-            // csharpier-ignore-end
-
-            int firstnumstates = FirstMaxVal.GetAsArray1D()[0];
-            int secondnumstates = SecondMaxVal.GetAsArray1D()[0];
-            //Console.WriteLine(firstnumstates);
-            //Console.WriteLine(secondnumstates);
-
-            using var JointBuffer = accelerate.Allocate2DDenseX<double>(
-                new Index2D(firstnumstates + 1, secondnumstates + 1)
-            );
-
-            using var FirstCountMap = accelerate.Allocate1D<double>(new Index1D(firstnumstates));
-            using var SecondCountMap = accelerate.Allocate1D<double>(new Index1D(secondnumstates));
-            //setBuffToValue2DKern(JointBuffer.Extent.ToIntIndex(), JointBuffer.View, 0.0);
-            // setBuffToValueDoubleKern(FirstBuffer.Extent.ToIntIndex(), FirstBuffer.View, 0.0);
-            // setBuffToValueDoubleKern(FirstBuffer.Extent.ToIntIndex(), FirstBuffer.View, 0.0);
-
-            BuildJointFreqKern(
-                SecondNormBuffer.Extent.ToIntIndex(),
-                FirstNormBuffer.View,
-                SecondNormBuffer.View,
-                JointBuffer.View
-            );
-            BuildFreqKern(SecondNormBuffer.Extent.ToIntIndex(), SecondNormBuffer.View, SecondCountMap.View);
-            BuildFreqKern(FirstNormBuffer.Extent.ToIntIndex(), FirstNormBuffer.View, FirstCountMap.View);
-
-            //CalcMIKern(JointBuffer.Extent.ToIntIndex(), JointBuffer.View, FirstCountMap.View,SecondCountMap.View, EntropyBuffer.View, firstVector.GetLength(0));
-
-            IndexedCalcMIKern(
-                SecondNormBuffer.Extent.ToIntIndex(),
-                JointBuffer.View,
-                FirstCountMap.View,
-                SecondCountMap.View,
-                FirstNormBuffer.View,
-                SecondNormBuffer.View,
-                EntropyBuffer.View,
-                firstVector.GetLength(0)
-            );
-
-            double answer = EntropyBuffer.GetAsArray1D()[0] / Math.Log(LOG_BASE);
-            accelerate.Dispose();
-            return answer;
-        }
-
-        public List<double> MulticalculateMutualInformation(List<double[]> arr)
-        {
-            using Accelerator accelerate = this.dev.CreateAccelerator(this.context);
-            using var FirstBuffer = accelerate.Allocate1D<double>(new Index1D(arr[0].GetLength(0)));
-            using var SecondBuffer = accelerate.Allocate1D<double>(new Index1D(arr[0].GetLength(0)));
-
-            using var FirstNormBuffer = accelerate.Allocate1D<double>(new Index1D(arr[0].GetLength(0)));
-            using var SecondNormBuffer = accelerate.Allocate1D<double>(new Index1D(arr[0].GetLength(0)));
-            using var FirstMaxVal = accelerate.Allocate1D<int>(new Index1D(1));
-            using var FirstMinVal = accelerate.Allocate1D<int>(new Index1D(1));
-
-            using var SecondMaxVal = accelerate.Allocate1D<int>(new Index1D(1));
-            using var SecondMinVal = accelerate.Allocate1D<int>(new Index1D(1));
-            using var EntropyBuffer = accelerate.Allocate1D<double>(new Index1D(1));
-            var JointBuffer = accelerate.Allocate2DDenseX<double>(new Index2D(1, 1));
-            var FirstCountMap = accelerate.Allocate1D<double>(new Index1D(1));
-            var SecondCountMap = accelerate.Allocate1D<double>(new Index1D(1));
-            int firstnumstates;
-            int secondnumstates;
-            double answer;
-
-            var GetMaxValKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>
-            >(GetMaxMinValKernal);
-            var InitMaxMinKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>
-            >(InitMaxMinKernel);
-            var normalizeArrayKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>
-            >(normalizeArrayKernel);
-
-            var BuildJointFreqKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView2D<double, Stride2D.DenseX>
-            >(BuildJointFreqKernel);
-
-            var setBuffToValue2DKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index2D,
-                ArrayView2D<double, Stride2D.DenseX>,
-                double
-            >(setBuffToValue2DKernal);
-
-            var CalcMIKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index2D,
-                ArrayView2D<double, Stride2D.DenseX>,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>,
-                int
-            >(CalcMIKernel);
-
-            var IndexedCalcMIKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView2D<double, Stride2D.DenseX>,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>,
-                int
-            >(IndexedCalcMIKernel);
-
-            var setBuffToValueDoubleKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                double
-            >(setBuffToValueDoubleKernal);
-
-            var BuildFreqKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>
-            >(BuildFreqAdjustedKernel);
-            List<double> MIanswers = new List<double>();
-
-            for (int i = 0; i < arr.Count; i++)
-            {
-                for (int j = 0; j < arr.Count; j++)
-                {
-                    if (i != j)
-                    {
-                        FirstBuffer.CopyFromCPU(arr[i]);
-                        SecondBuffer.CopyFromCPU(arr[j]);
-                        InitMaxMinKern(
-                            FirstMaxVal.Extent.ToIntIndex(),
-                            FirstBuffer.View,
-                            FirstMaxVal.View,
-                            FirstMinVal.View
-                        );
-                        GetMaxValKern(
-                            FirstBuffer.Extent.ToIntIndex(),
-                            FirstBuffer.View,
-                            FirstMaxVal.View,
-                            FirstMinVal.View
-                        );
-                        //Console.WriteLine(FirstMaxVal.GetAsArray1D()[0]);
-                        InitMaxMinKern(
-                            SecondMaxVal.Extent.ToIntIndex(),
-                            SecondBuffer.View,
-                            SecondMaxVal.View,
-                            SecondMinVal.View
-                        );
-                        GetMaxValKern(
-                            SecondBuffer.Extent.ToIntIndex(),
-                            SecondBuffer.View,
-                            SecondMaxVal.View,
-                            SecondMinVal.View
-                        );
-
-                        normalizeArrayKern(
-                            FirstBuffer.Extent.ToIntIndex(),
-                            FirstBuffer.View,
-                            FirstNormBuffer.View,
-                            FirstMinVal
-                        );
-                        normalizeArrayKern(
-                            SecondBuffer.Extent.ToIntIndex(),
-                            SecondBuffer.View,
-                            SecondNormBuffer.View,
-                            SecondMinVal
-                        );
-
-                        firstnumstates = FirstMaxVal.GetAsArray1D()[0];
-                        secondnumstates = SecondMaxVal.GetAsArray1D()[0];
-                        //Console.WriteLine(firstnumstates);
-                        //Console.WriteLine(secondnumstates);
-
-                        JointBuffer = accelerate.Allocate2DDenseX<double>(
-                            new Index2D(firstnumstates + 1, secondnumstates + 1)
-                        );
-
-                        FirstCountMap = accelerate.Allocate1D<double>(new Index1D(firstnumstates));
-                        SecondCountMap = accelerate.Allocate1D<double>(new Index1D(secondnumstates));
-                        //setBuffToValue2DKern(JointBuffer.Extent.ToIntIndex(), JointBuffer.View, 0.0);
-                        // setBuffToValueDoubleKern(FirstBuffer.Extent.ToIntIndex(), FirstBuffer.View, 0.0);
-                        // setBuffToValueDoubleKern(FirstBuffer.Extent.ToIntIndex(), FirstBuffer.View, 0.0);
-                        setBuffToValueDoubleKern(FirstCountMap.Extent.ToIntIndex(), FirstCountMap.View, 0.0);
-                        setBuffToValueDoubleKern(FirstCountMap.Extent.ToIntIndex(), FirstCountMap.View, 0.0);
-                        BuildJointFreqKern(
-                            SecondNormBuffer.Extent.ToIntIndex(),
-                            FirstNormBuffer.View,
-                            SecondNormBuffer.View,
-                            JointBuffer.View
-                        );
-                        BuildFreqKern(SecondNormBuffer.Extent.ToIntIndex(), SecondNormBuffer.View, SecondCountMap.View);
-                        BuildFreqKern(FirstNormBuffer.Extent.ToIntIndex(), FirstNormBuffer.View, FirstCountMap.View);
-
-                        //CalcMIKern(JointBuffer.Extent.ToIntIndex(), JointBuffer.View, FirstCountMap.View,SecondCountMap.View, EntropyBuffer.View, firstVector.GetLength(0));
-                        setBuffToValueDoubleKern(EntropyBuffer.Extent.ToIntIndex(), EntropyBuffer.View, 0.0);
-                        IndexedCalcMIKern(
-                            SecondNormBuffer.Extent.ToIntIndex(),
-                            JointBuffer.View,
-                            FirstCountMap.View,
-                            SecondCountMap.View,
-                            FirstNormBuffer.View,
-                            SecondNormBuffer.View,
-                            EntropyBuffer.View,
-                            arr[0].GetLength(0)
-                        );
-
-                        answer = EntropyBuffer.GetAsArray1D()[0] / Math.Log(LOG_BASE);
-                        MIanswers.Add(answer);
-                    }
-                }
-            }
-            JointBuffer.Dispose();
-            FirstCountMap.Dispose();
-            SecondCountMap.Dispose();
-            accelerate.Dispose();
-            return MIanswers;
-        }
-
-        public double calculateConditionalMutualInformation(
-            double[] firstVector,
-            double[] secondVector,
-            double[] conditionVector
-        )
-        {
-            //FirstCondEntropy (secondvector, conditionvector)
-            //SecondCondEntropy (secondvector, mergedvector)
-            Stopwatch watch = new Stopwatch();
-            watch.Start();
-            using Accelerator accelerate = this.dev.CreateAccelerator(this.context);
-            using var FirstBuffer = accelerate.Allocate1D<double>(new Index1D(firstVector.GetLength(0)));
-            using var SecondBuffer = accelerate.Allocate1D<double>(new Index1D(secondVector.GetLength(0)));
-            using var CondBuffer = accelerate.Allocate1D<double>(new Index1D(conditionVector.GetLength(0)));
-            var testIndex = new LongIndex2D(10000, 1000000);
-            using var FirstNormBuffer = accelerate.Allocate1D<double>(new Index1D(firstVector.GetLength(0)));
-            using var SecondNormBuffer = accelerate.Allocate1D<double>(new Index1D(secondVector.GetLength(0)));
-            using var CondNormBuffer = accelerate.Allocate1D<double>(new Index1D(conditionVector.GetLength(0)));
-
-            using var mergedBuffer = accelerate.Allocate1D<double>(new Index1D(firstVector.GetLength(0)));
-            //var AdjMergeBuffer = accelerate.Allocate1D<double>(new Index1D(firstVector.GetLength(0)));
-
-            using var MergeNormBuffer = accelerate.Allocate1D<double>(new Index1D(firstVector.GetLength(0)));
-
-            FirstBuffer.CopyFromCPU(firstVector);
-            SecondBuffer.CopyFromCPU(secondVector);
-            CondBuffer.CopyFromCPU(conditionVector);
-
-            using var FirstMaxVal = accelerate.Allocate1D<int>(new Index1D(1));
-            using var FirstMinVal = accelerate.Allocate1D<int>(new Index1D(1));
-
-            using var SecondMaxVal = accelerate.Allocate1D<int>(new Index1D(1));
-            using var SecondMinVal = accelerate.Allocate1D<int>(new Index1D(1));
-
-            using var CondMaxVal = accelerate.Allocate1D<int>(new Index1D(1));
-            using var CondMinVal = accelerate.Allocate1D<int>(new Index1D(1));
-
-            using var MergeMaxVal = accelerate.Allocate1D<int>(new Index1D(1));
-            using var MergeMinVal = accelerate.Allocate1D<int>(new Index1D(1));
-
-            using var PreMergeMaxVal = accelerate.Allocate1D<int>(new Index1D(1));
-            using var PreMergeMinVal = accelerate.Allocate1D<int>(new Index1D(1));
-
-            using var EntropyBuffer1 = accelerate.Allocate1D<double>(new Index1D(1));
-            using var EntropyBuffer2 = accelerate.Allocate1D<double>(new Index1D(1));
-
-            using var StateCount = accelerate.Allocate1D<int>(new Index1D(1));
-
-            using var CountNonNaN = accelerate.Allocate1D<double>(new Index1D(1));
-
-            var GetMaxValKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>
-            >(TestGetMaxMinValKernal);
-            var InitMaxMinKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>
-            >(InitMaxMinKernel);
-            var normalizeArrayKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>
-            >(normalizeArrayKernel);
-            var mergeArraysKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<int, Stride1D.Dense>,
-                int,
-                int
-            >(mergeArraysAdjKernel);
-            var BuildJointFreqKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView2D<double, Stride2D.DenseX>
-            >(BuildJointFreqKernel);
-
-            var setBuffToValue2DKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index2D,
-                ArrayView2D<double, Stride2D.DenseX>,
-                double
-            >(setBuffToValue2DKernal);
-
-            var CalcConditionalEntropyKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index2D,
-                ArrayView2D<double, Stride2D.DenseX>,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>,
-                int
-            >(CalcConditionalEntropyKernel);
-
-            var IndexedCalcConditionalEntropyKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView2D<double, Stride2D.DenseX>,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>,
-                int
-            >(IndexedCalcConditionalEntropyKernel);
-
-            var setBuffToValueDoubleKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                double
-            >(setBuffToValueDoubleKernal);
-
-            var setBuffToValueKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<int, Stride1D.Dense>,
-                int
-            >(setBuffToValueKernal);
-
-            var BuildFreqKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>
-            >(BuildFreqAdjKernel);
-
-            var RefactorPart1Kern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                int
-            >(refactorKernal);
-
-            var refactorPart1Phase1Kern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>
-            >(refactorPart1Phase1Kernal);
-
-            var refactorPart1Phase2Kern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>
-            >(refactorPart1Phase2Kernal);
-
-            var refactorPart2Phase1Kern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>
-            >(refactorPart2Phase1Kernal);
-
-            var refactorPart2Phase2Kern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>
-            >(refactorPart2Phase2Kernal);
-
-            var countnonNaNKern = accelerate.LoadAutoGroupedStreamKernel<
-                Index1D,
-                ArrayView1D<double, Stride1D.Dense>,
-                ArrayView1D<double, Stride1D.Dense>
-            >(countnonNaN);
-
-            // watch.Stop();
-            // Console.WriteLine("Initialization: ");
-            // Console.WriteLine(watch.ElapsedMilliseconds);
-            // watch.Reset();
-            // accelerate.DefaultStream.Synchronize();
-            // watch.Start();
-            InitMaxMinKern(FirstMaxVal.Extent.ToIntIndex(), FirstBuffer.View, FirstMaxVal.View, FirstMinVal.View);
-            GetMaxValKern(FirstBuffer.Extent.ToIntIndex(), FirstBuffer.View, FirstMaxVal.View, FirstMinVal.View);
-
-            InitMaxMinKern(SecondMaxVal.Extent.ToIntIndex(), SecondBuffer.View, SecondMaxVal.View, SecondMinVal.View);
-            GetMaxValKern(SecondBuffer.Extent.ToIntIndex(), SecondBuffer.View, SecondMaxVal.View, SecondMinVal.View);
-
-            InitMaxMinKern(CondMaxVal.Extent.ToIntIndex(), CondBuffer.View, CondMaxVal.View, CondMinVal.View);
-            GetMaxValKern(CondBuffer.Extent.ToIntIndex(), CondBuffer.View, CondMaxVal.View, CondMinVal.View);
-
-            normalizeArrayKern(
-                FirstBuffer.Extent.ToIntIndex(),
-                FirstBuffer.View,
-                FirstNormBuffer.View,
-                FirstMinVal.View
-            );
-            normalizeArrayKern(
-                SecondBuffer.Extent.ToIntIndex(),
-                SecondBuffer.View,
-                SecondNormBuffer.View,
-                SecondMinVal.View
-            );
-            normalizeArrayKern(CondBuffer.Extent.ToIntIndex(), CondBuffer.View, CondNormBuffer.View, CondMinVal.View);
-            // accelerate.DefaultStream.Synchronize();
-
-            // watch.Stop();
-            // //RefactorPart1TestKern(testIndex, FirstBuffer.View, SecondBuffer.View);
-            // Console.WriteLine("MaxMin ETC: ");
-
-            // Console.WriteLine(watch.ElapsedMilliseconds);
-            // watch.Reset();
-
-            // accelerate.DefaultStream.Synchronize();
-            // watch.Start();
-
-            // csharpier-ignore-start
-            // Reusing these variables from before because we don't need the original values anymore now that everything is normalized
-            InitMaxMinKern(FirstMaxVal.Extent.ToIntIndex(), FirstNormBuffer.View, FirstMaxVal.View, FirstMinVal.View);
-            GetMaxValKern(FirstNormBuffer.Extent.ToIntIndex(), FirstNormBuffer.View, FirstMaxVal.View, FirstMinVal.View);
-            InitMaxMinKern(SecondMaxVal.Extent.ToIntIndex(), SecondNormBuffer.View, SecondMaxVal.View, SecondMinVal.View);
-            GetMaxValKern(SecondNormBuffer.Extent.ToIntIndex(), SecondNormBuffer.View, SecondMaxVal.View, SecondMinVal.View);
-            InitMaxMinKern(CondMaxVal.Extent.ToIntIndex(), CondNormBuffer.View, CondMaxVal.View, CondMinVal.View);
-            GetMaxValKern(CondNormBuffer.Extent.ToIntIndex(), CondNormBuffer.View, CondMaxVal.View, CondMinVal.View);
-            // csharpier-ignore-end
-
-            int firstnumstates = FirstMaxVal.GetAsArray1D()[0];
-            int secondnumstates = SecondMaxVal.GetAsArray1D()[0];
-            int condnumstates = CondMaxVal.GetAsArray1D()[0];
-
-            using var StateMap = accelerate.Allocate1D<int>(new Index1D((firstnumstates + 1) * (secondnumstates + 1)));
-            //outputVector = mergedBuffer.GetAsArray1D();
-
-            setBuffToValueKern(StateCount.Extent.ToIntIndex(), StateCount.View, 1);
-            //print1d(FirstNormBuffer.GetAsArray1D());
-            //print1d(SecondNormBuffer.GetAsArray1D());
-            //Console.ReadLine();
-            mergeArraysKern(
-                mergedBuffer.Extent.ToIntIndex(),
-                FirstNormBuffer.View,
-                SecondNormBuffer.View,
-                StateMap.View,
-                mergedBuffer.View,
-                StateCount.View,
-                firstnumstates,
-                FirstNormBuffer.Extent.ToIntIndex().X
-            );
-            // mergeArraysKern(StateCount.Extent.ToIntIndex(), FirstNormBuffer.View, SecondNormBuffer.View, StateMap.View, mergedBuffer.View,StateCount.View, firstnumstates, FirstNormBuffer.Extent.ToIntIndex().X);
-
-            //outputVector = mergedBuffer.GetAsArray1D();
-            // print1d(outputVector);
-            // Console.WriteLine("InHere");
-            // print1d(SecondNormBuffer.GetAsArray1D());
-            // print1d(FirstNormBuffer.GetAsArray1D());
-            //print1d(mergedBuffer.GetAsArray1D());
-            // Console.ReadLine();
-            InitMaxMinKern(
-                PreMergeMaxVal.Extent.ToIntIndex(),
-                mergedBuffer.View,
-                PreMergeMaxVal.View,
-                PreMergeMinVal.View
-            );
-            GetMaxValKern(
-                mergedBuffer.Extent.ToIntIndex(),
-                mergedBuffer.View,
-                PreMergeMaxVal.View,
-                PreMergeMinVal.View
-            );
-            //accelerate.DefaultStream.Synchronize();
-            int premergenumstates = PreMergeMaxVal.GetAsArray1D()[0];
-            using var TempBuffer = accelerate.Allocate1D<double>(new Index1D(premergenumstates + 1));
-            //var Temp2Buffer = accelerate.Allocate1D<double>(new Index1D(premergenumstates +1));
-
-            using var HolderBuffer = accelerate.Allocate1D<double>(new Index1D(1));
-            //accelerate.DefaultStream.Synchronize();
-            // print1d(PreMergeMinVal.GetAsArray1D());
-            // Console.ReadLine();
-            // Console.WriteLine("HERHEREHERHERHEHREREH");
-
-            // watch.Stop();
-
-            //Console.WriteLine("Merge Arrs:");
-            //Console.WriteLine(watch.ElapsedMilliseconds);
-            //accelerate.DefaultStream.Synchronize();
-
-            //watch.Reset();
-            //watch.Start();
-
-            ///WILL PUT THIS IN THE GPU
-            //RefactorPart1Kern(mergedBuffer.Extent.ToIntIndex(), mergedBuffer.View, firstVector.GetLength(0));
-            refactorPart1Phase1Kern(mergedBuffer.Extent.ToIntIndex(), mergedBuffer.View, TempBuffer.View);
-            //Console.WriteLine(premergenumstates);
-            //print1d(PreMergeMaxVal.GetAsArray1D());
-            //Console.ReadLine();
-            //print1d(TempBuffer.GetAsArray1D());
-
-            refactorPart1Phase2Kern(mergedBuffer.Extent.ToIntIndex(), mergedBuffer.View, TempBuffer.View);
-            //accelerate.DefaultStream.Synchronize();
-            //print1d(mergedBuffer.GetAsArray1D());
-            watch.Stop();
-            //print1d(mergedBuffer.GetAsArray1D());
-            //Console.ReadLine();
-
-            // Console.WriteLine("Part 1:");
-            // Console.WriteLine(watch.ElapsedMilliseconds);
-            // Console.ReadLine();
-
-            watch.Reset();
-            //print1d(mergedBuffer.GetAsArray1D());
-            //Console.WriteLine("Testtiong");
-            //Console.ReadLine();
-            watch.Start();
-            setBuffToValueDoubleKern(TempBuffer.Extent.ToIntIndex(), TempBuffer.View, 1.0);
-
-            setBuffToValueDoubleKern(HolderBuffer.Extent.ToIntIndex(), HolderBuffer.View, 1.0);
-
-            refactorPart2Phase1Kern(
-                mergedBuffer.Extent.ToIntIndex(),
-                mergedBuffer.View,
-                TempBuffer.View,
-                HolderBuffer.View
-            );
-            //print1d(TempBuffer.GetAsArray1D());
-            //print1d(HolderBuffer.GetAsArray1D());
-            //Console.ReadLine();
-
-            refactorPart2Phase2Kern(mergedBuffer.Extent.ToIntIndex(), mergedBuffer.View, TempBuffer.View);
-
-            //print1d(mergedBuffer.GetAsArray1D());
-            //Console.ReadLine();
-            //AdjMergeBuffer.CopyFromCPU(refactorToMinimizeSize(mergedBuffer.GetAsArray1D()));
-
-            //print1d(mergedBuffer.GetAsArray1D());
-            //accelerate.DefaultStream.Synchronize();
-
-            watch.Stop();
-
-            // Console.WriteLine("Refactor Arrs Part2:");
-
-            //Console.WriteLine(watch.ElapsedMilliseconds);
-            watch.Reset();
-            InitMaxMinKern(MergeMaxVal.Extent.ToIntIndex(), mergedBuffer.View, MergeMaxVal.View, MergeMinVal.View);
-            GetMaxValKern(mergedBuffer.Extent.ToIntIndex(), mergedBuffer.View, MergeMaxVal.View, MergeMinVal.View);
-            //print1d(MergeMaxVal.GetAsArray1D());
-
-            // print1d(MergeMaxVal.GetAsArray1D());
-            // print1d(MergeMinVal.GetAsArray1D());
-            // print1d(mergedBuffer.GetAsArray1D());
-            // Console.ReadLine();
-
-            normalizeArrayKern(
-                mergedBuffer.Extent.ToIntIndex(),
-                mergedBuffer.View,
-                MergeNormBuffer.View,
-                MergeMinVal.View
-            );
-
-            int mergenumstates = MergeMaxVal.GetAsArray1D()[0];
-
-            setBuffToValueDoubleKern(CountNonNaN.Extent.ToIntIndex(), CountNonNaN.View, 1.0);
-            countnonNaNKern(mergedBuffer.Extent.ToIntIndex(), mergedBuffer.View, CountNonNaN.View);
-
-            // Console.WriteLine("NumStates");
-            // Console.WriteLine(secondnumstates);
-
-            // Console.WriteLine(condnumstates);
-
-            // Console.WriteLine(mergenumstates);
-            // Console.WriteLine(MergeMinVal.GetAsArray1D()[0]);
-            // Console.WriteLine("PreMergeNumStates");
-            // Console.WriteLine(premergenumstates);
-
-            // Console.WriteLine(CountNonNaN.GetAsArray1D()[0]);
-
-            //print1d(mergedBuffer.GetAsArray1D());
-            //Console.ReadLine();
-
-            using var JointBuffer1 = accelerate.Allocate2DDenseX<double>(
-                new Index2D(secondnumstates + 1, condnumstates + 1)
-            );
-            using var JointBuffer2 = accelerate.Allocate2DDenseX<double>(
-                new Index2D(secondnumstates + 1, mergenumstates + 1)
-            );
-
-            //EntropyBuffer1.GetAsArray1D();
-
-            //var SecondCountMap = accelerate.Allocate1D<double>(new Index1D(secondnumstates));
-            using var CondCountMap = accelerate.Allocate1D<double>(new Index1D(condnumstates));
-            using var MergeCountMap = accelerate.Allocate1D<double>(new Index1D(mergenumstates));
-
-            // print1d(CondNormBuffer.GetAsArray1D());
-            // print1d(MergeNormBuffer.GetAsArray1D());
-
-            BuildJointFreqKern(
-                CondNormBuffer.Extent.ToIntIndex(),
-                SecondNormBuffer.View,
-                CondNormBuffer.View,
-                JointBuffer1.View
-            );
-            //EntropyBuffer1.GetAsArray1D();
-            //EntropyBuffer1.GetAsArray1D();
-
-            BuildJointFreqKern(
-                MergeNormBuffer.Extent.ToIntIndex(),
-                SecondNormBuffer.View,
-                MergeNormBuffer.View,
-                JointBuffer2.View
-            );
-            //EntropyBuffer1.GetAsArray1D();
-
-            //EntropyBuffer1.GetAsArray1D();
-
-            BuildFreqKern(CondNormBuffer.Extent.ToIntIndex(), CondNormBuffer.View, CondCountMap.View);
-            //EntropyBuffer1.GetAsArray1D();
-            //print1d(MergeNormBuffer.GetAsArray1D());
-            BuildFreqKern(MergeNormBuffer.Extent.ToIntIndex(), MergeNormBuffer.View, MergeCountMap.View);
-            //EntropyBuffer1.GetAsArray1D();
-
-            //EntropyBuffer1.GetAsArray1D();
-            IndexedCalcConditionalEntropyKern(
-                CondNormBuffer.Extent.ToIntIndex(),
-                JointBuffer1.View,
-                CondCountMap.View,
-                SecondNormBuffer.View,
-                CondNormBuffer.View,
-                EntropyBuffer1.View,
-                secondVector.GetLength(0)
-            );
-            IndexedCalcConditionalEntropyKern(
-                MergeNormBuffer.Extent.ToIntIndex(),
-                JointBuffer2.View,
-                MergeCountMap.View,
-                SecondNormBuffer.View,
-                MergeNormBuffer.View,
-                EntropyBuffer2.View,
-                secondVector.GetLength(0)
-            );
-
-            double ent1 = EntropyBuffer1.GetAsArray1D()[0] / Math.Log(LOG_BASE);
-
-            double ent2 = EntropyBuffer2.GetAsArray1D()[0] / Math.Log(LOG_BASE);
-
-            return ent1 - ent2;
-        }
-
-        public double ConditiponalMIHolder(double[] firstVector, double[] secondVector, double[] conditionVector)
-        {
-            double[] mergedVector = new double[firstVector.GetLength(0)];
-
-            mergeArrays(firstVector, conditionVector, ref mergedVector);
-
-            //Secondvector = first vector
-            // conditionvector = second vector
-            double firstCondEnt = calculateConditionalEntropy(secondVector, conditionVector);
-            //print1d(refactorToMinimizeSize(mergedVector));
-
-            //second vector = first vector
-            // merged vector = second vector
-            double secondCondEnt = calculateConditionalEntropy(secondVector, refactorToMinimizeSize(mergedVector));
-
-            return firstCondEnt - secondCondEnt;
-        }
-
-        static void replaceNaNKernal(Index1D index, ArrayView1D<double, Stride1D.Dense> inputView)
+        // public double mergeArrays(double[] firstVector, double[] secondVector, ref double[] outputVector)
+        // {
+        //     using Accelerator accelerate = this.dev.CreateAccelerator(this.context);
+        //     using var FirstBuffer = accelerate.Allocate1D<double>(new Index1D(firstVector.GetLength(0)));
+        //     using var SecondBuffer = accelerate.Allocate1D<double>(new Index1D(secondVector.GetLength(0)));
+
+        //     using var FirstNormBuffer = accelerate.Allocate1D<double>(new Index1D(firstVector.GetLength(0)));
+        //     using var SecondNormBuffer = accelerate.Allocate1D<double>(new Index1D(secondVector.GetLength(0)));
+
+        //     using var outputBuffer = accelerate.Allocate1D<double>(new Index1D(outputVector.GetLength(0)));
+
+        //     FirstBuffer.CopyFromCPU(firstVector);
+        //     SecondBuffer.CopyFromCPU(secondVector);
+
+        //     using var FirstMaxVal = accelerate.Allocate1D<int>(new Index1D(1));
+        //     using var FirstMinVal = accelerate.Allocate1D<int>(new Index1D(1));
+
+        //     using var SecondMaxVal = accelerate.Allocate1D<int>(new Index1D(1));
+        //     using var SecondMinVal = accelerate.Allocate1D<int>(new Index1D(1));
+
+        //     using var StateCount = accelerate.Allocate1D<int>(new Index1D(1));
+
+        //     var GetMaxValKern = accelerate.LoadAutoGroupedStreamKernel<
+        //         Index1D,
+        //         ArrayView1D<double, Stride1D.Dense>,
+        //         ArrayView1D<int, Stride1D.Dense>,
+        //         ArrayView1D<int, Stride1D.Dense>
+        //     >(GetMaxMinValKernal);
+        //     var InitMaxMinKern = accelerate.LoadAutoGroupedStreamKernel<
+        //         Index1D,
+        //         ArrayView1D<double, Stride1D.Dense>,
+        //         ArrayView1D<int, Stride1D.Dense>,
+        //         ArrayView1D<int, Stride1D.Dense>
+        //     >(InitMaxMinKernel);
+        //     var normalizeArrayKern = accelerate.LoadAutoGroupedStreamKernel<
+        //         Index1D,
+        //         ArrayView1D<double, Stride1D.Dense>,
+        //         ArrayView1D<double, Stride1D.Dense>,
+        //         ArrayView1D<int, Stride1D.Dense>
+        //     >(normalizeArrayKernel);
+        //     var mergeArraysKern = accelerate.LoadAutoGroupedStreamKernel<
+        //         Index1D,
+        //         ArrayView1D<double, Stride1D.Dense>,
+        //         ArrayView1D<double, Stride1D.Dense>,
+        //         ArrayView1D<int, Stride1D.Dense>,
+        //         ArrayView1D<double, Stride1D.Dense>,
+        //         ArrayView1D<int, Stride1D.Dense>,
+        //         int,
+        //         int
+        //     >(mergeArraysKernel);
+
+        //     InitMaxMinKern(FirstMaxVal.Extent.ToIntIndex(), FirstBuffer.View, FirstMaxVal.View, FirstMinVal.View);
+        //     GetMaxValKern(FirstBuffer.Extent.ToIntIndex(), FirstBuffer.View, FirstMaxVal.View, FirstMinVal.View);
+
+        //     InitMaxMinKern(SecondMaxVal.Extent.ToIntIndex(), SecondBuffer.View, SecondMaxVal.View, SecondMinVal.View);
+        //     GetMaxValKern(SecondBuffer.Extent.ToIntIndex(), SecondBuffer.View, SecondMaxVal.View, SecondMinVal.View);
+
+        //     normalizeArrayKern(FirstBuffer.Extent.ToIntIndex(), FirstBuffer.View, FirstNormBuffer.View, FirstMinVal);
+        //     normalizeArrayKern(
+        //         SecondBuffer.Extent.ToIntIndex(),
+        //         SecondBuffer.View,
+        //         SecondNormBuffer.View,
+        //         SecondMinVal
+        //     );
+        //     int firstnumstates = FirstMaxVal.GetAsArray1D()[0];
+        //     int secondnumstates = SecondMaxVal.GetAsArray1D()[0];
+
+        //     using var StateMap = accelerate.Allocate1D<int>(new Index1D((firstnumstates + 1) * (secondnumstates + 1)));
+        //     //outputVector = outputBuffer.GetAsArray1D();
+        //     mergeArraysKern(
+        //         StateCount.Extent.ToIntIndex(),
+        //         FirstNormBuffer.View,
+        //         SecondNormBuffer.View,
+        //         StateMap.View,
+        //         outputBuffer.View,
+        //         StateCount.View,
+        //         firstnumstates,
+        //         FirstNormBuffer.Extent.ToIntIndex().X
+        //     );
+        //     outputVector = outputBuffer.GetAsArray1D();
+        //     // print1d(outputVector);
+        //     double answer = StateCount.GetAsArray1D()[0];
+        //     accelerate.Dispose();
+        //     return answer;
+        // }
+
+        protected static void replaceNaNKernal(Index1D index, ArrayView1D<double, Stride1D.Dense> inputView)
         {
             if (Double.IsNaN(inputView[index]))
             {
@@ -1829,97 +240,11 @@ namespace CSMI
             }
         }
 
-        static void refactorKernal(Index1D index, ArrayView1D<double, Stride1D.Dense> inputView, int length)
-        {
-            int holder = 0;
-
-            for (int i = 0; i < length; i++)
-            {
-                if (inputView[index] == inputView[new Index1D(i)] && index.X != i)
-                {
-                    holder = 1;
-                }
-            }
-            if (holder == 0)
-            {
-                inputView[index] = Double.NaN;
-            }
-        }
-
         // static void fixFirst(Index1D index,
         //     ArrayView1D<double, Stride1D.Dense> inputView){
         //     if(Double.IsNaN(inputView[index]))
         // }
-        static void refactorPart1Phase1Kernal(
-            Index1D index,
-            ArrayView1D<double, Stride1D.Dense> inputView,
-            ArrayView1D<double, Stride1D.Dense> holderView
-        )
-        {
-            if (!Double.IsNaN(inputView[index]) && inputView[index] > 1000)
-            {
-                Atomic.Add(ref holderView[new Index1D((int)inputView[index])], 1.0);
-            }
-        }
 
-        static void refactorPart1Phase2Kernal(
-            Index1D index,
-            ArrayView1D<double, Stride1D.Dense> inputView,
-            ArrayView1D<double, Stride1D.Dense> holderView
-        )
-        {
-            if (holderView[new Index1D((int)inputView[index])] <= 1.0)
-            {
-                inputView[index] = Double.NaN;
-            }
-        }
-
-        static void refactorPart2Phase1Kernal(
-            Index1D index,
-            ArrayView1D<double, Stride1D.Dense> inputView,
-            ArrayView1D<double, Stride1D.Dense> holderView,
-            ArrayView1D<double, Stride1D.Dense> sharedmem
-        )
-        {
-            if (!Double.IsNaN(inputView[index]))
-            {
-                // //double x = Atomic.Add(ref holderView[(int)inputView[index]], Atomic.Add(ref sharedmem[new Index1D(0)], 1.0));
-                // if(holderView[(int)inputView[index]] > 1.0){
-                //     //holderView[(int)inputView[index]] = x;
-                //     //Atomic.Add(ref holderView[(int)inputView[index]], sharedmem[new Index1D(0)]);
-                // }
-                // else{
-                holderView[(int)inputView[index]] = Atomic.Add(ref sharedmem[new Index1D(0)], 1.0);
-            }
-        }
-
-        static void countnonNaN(
-            Index1D index,
-            ArrayView1D<double, Stride1D.Dense> inputView,
-            ArrayView1D<double, Stride1D.Dense> output
-        )
-        {
-            if (!Double.IsNaN(inputView[index]))
-            {
-                Atomic.Add(ref output[new Index1D(0)], 1.0);
-            }
-        }
-
-        static void refactorPart2Phase2Kernal(
-            Index1D index,
-            ArrayView1D<double, Stride1D.Dense> inputView,
-            ArrayView1D<double, Stride1D.Dense> holderView
-        )
-        {
-            if (!Double.IsNaN(inputView[index]))
-            {
-                inputView[index] = holderView[(int)inputView[index]];
-            }
-            else if (index.X == 0 && Double.IsNaN(inputView[index]))
-            {
-                inputView[new Index1D(0)] = 1.0;
-            }
-        }
 
         // static void refactorPart2Kernal(Index1D index,
         //     ArrayView1D<double, Stride1D.Dense> inputView,
@@ -1954,7 +279,7 @@ namespace CSMI
         //     }
         // }
 
-        static void mergeArraysKernel(
+        protected static void mergeArraysKernel(
             Index1D index,
             ArrayView1D<double, Stride1D.Dense> FirstNormView,
             ArrayView1D<double, Stride1D.Dense> SecondNormView,
@@ -1980,305 +305,13 @@ namespace CSMI
             SCount[index] = statecount;
         }
 
-        static void mergeArraysAdjKernel(
-            Index1D index,
-            ArrayView1D<double, Stride1D.Dense> FirstNormView,
-            ArrayView1D<double, Stride1D.Dense> SecondNormView,
-            ArrayView1D<int, Stride1D.Dense> StateMap,
-            ArrayView1D<double, Stride1D.Dense> OutputView,
-            ArrayView1D<int, Stride1D.Dense> SCount,
-            int firstnumstates,
-            int length
-        )
-        {
-            int curindex;
-            //int statecount = 1;
-            //Probably doesnt work
-            curindex = (int)FirstNormView[index] + ((int)SecondNormView[index] * firstnumstates);
-            // if(Atomic.Add(ref StateMap[new Index1D(curindex)], 1) == 0){
-            //     StateMap[new Index1D(curindex)] = Atomic.Add(ref SCount[new Index1D(0)], 1);
-            // }
-            // else{
-            //     Atomic.Add(ref StateMap[new Index1D(curindex)], -1);
-            // }
-            if (StateMap[new Index1D(curindex)] == 0)
-            {
-                StateMap[new Index1D(curindex)] = Atomic.Add(ref SCount[new Index1D(0)], 1);
-            }
-            OutputView[index] = (double)StateMap[curindex];
-        }
-
-        static void CalcEntropyKernel(
-            Index1D index,
-            ArrayView1D<double, Stride1D.Dense> FreqView,
-            ArrayView1D<double, Stride1D.Dense> entropy,
-            int length
-        )
-        {
-            double val;
-            if (FreqView[index] > 0)
-            {
-                val = -1 * (FreqView[index] / (double)length) * Math.Log(FreqView[index] / (double)length);
-                Atomic.Add(ref entropy[new Index1D(0)], val);
-            }
-        }
-
-        static void CalcConditionalEntropyKernel(
-            Index2D index,
-            ArrayView2D<double, Stride2D.DenseX> FreqView,
-            ArrayView1D<double, Stride1D.Dense> condView,
-            ArrayView1D<double, Stride1D.Dense> entropy,
-            int length
-        )
-        {
-            double val;
-            if (FreqView[index] > 0 && condView[index.Y] > 0)
-            {
-                val =
-                    -1
-                    * (
-                        (FreqView[index] / (double)length)
-                        * Math.Log((FreqView[index] / (double)length) / (condView[index.Y] / (double)length))
-                    );
-                Atomic.Add(ref entropy[new Index1D(0)], val);
-            }
-        }
-
-        static void IndexedCalcConditionalEntropyKernel(
-            Index1D index,
-            ArrayView2D<double, Stride2D.DenseX> FreqView,
-            ArrayView1D<double, Stride1D.Dense> condView,
-            ArrayView1D<double, Stride1D.Dense> FirstView,
-            ArrayView1D<double, Stride1D.Dense> SecondView,
-            ArrayView1D<double, Stride1D.Dense> entropy,
-            int length
-        )
-        {
-            double val;
-            if (!(Double.IsNaN(FirstView[index])) && !(Double.IsNaN(SecondView[index])))
-            {
-                Index2D ind2 = new Index2D((int)FirstView[index], (int)SecondView[index]);
-
-                if (FreqView[ind2] > 0 && condView[ind2.Y] > 0)
-                {
-                    val =
-                        -1
-                        * (
-                            (FreqView[ind2] / (double)length)
-                            * Math.Log((FreqView[ind2] / (double)length) / (condView[ind2.Y] / (double)length))
-                        );
-                    val = val / FreqView[ind2];
-                    Atomic.Add(ref entropy[new Index1D(0)], val);
-                }
-            }
-            else
-            {
-                val = -1 * ((1 / (double)length) * Math.Log((1 / (double)length) / (1 / (double)length)));
-
-                Atomic.Add(ref entropy[new Index1D(0)], val);
-            }
-        }
-
-        static void CalcMIKernel(
-            Index2D index,
-            ArrayView2D<double, Stride2D.DenseX> FreqView,
-            ArrayView1D<double, Stride1D.Dense> first,
-            ArrayView1D<double, Stride1D.Dense> second,
-            ArrayView1D<double, Stride1D.Dense> entropy,
-            int length
-        )
-        {
-            double val;
-            if (FreqView[index] > 0 && first[index.X] > 0 && second[index.Y] > 0)
-            {
-                val = (
-                    (FreqView[index] / (double)length)
-                    * Math.Log(
-                        ((FreqView[index] / (double)length) / (first[index.X] / (double)length))
-                            / (second[index.Y] / (double)length)
-                    )
-                );
-                Atomic.Add(ref entropy[new Index1D(0)], val);
-            }
-        }
-
-        static void IndexedCalcMIKernel(
-            Index1D index,
-            ArrayView2D<double, Stride2D.DenseX> FreqView,
-            ArrayView1D<double, Stride1D.Dense> first,
-            ArrayView1D<double, Stride1D.Dense> second,
-            ArrayView1D<double, Stride1D.Dense> FirstView,
-            ArrayView1D<double, Stride1D.Dense> SecondView,
-            ArrayView1D<double, Stride1D.Dense> entropy,
-            int length
-        )
-        {
-            double val;
-            Index2D ind2 = new Index2D((int)FirstView[index], (int)SecondView[index]);
-            if (FreqView[ind2] > 0 && first[(int)FirstView[index]] > 0 && second[(int)SecondView[index]] > 0)
-            {
-                // csharpier-ignore
-                val = ((FreqView[ind2] / (double)length) * Math.Log(((FreqView[ind2] / (double)length) / (first[ind2.X] / (double)length)) / (second[ind2.Y] / (double)length)));
-                val = val / FreqView[ind2];
-                Atomic.Add(ref entropy[new Index1D(0)], val);
-            }
-        }
-
-        static void CalcJointEntropyKernel(
-            Index2D index,
-            ArrayView2D<double, Stride2D.DenseX> FreqView,
-            ArrayView1D<double, Stride1D.Dense> entropy,
-            int length
-        )
-        {
-            double val;
-            if (FreqView[index] > 0)
-            {
-                val = -1 * (FreqView[index] / (double)length) * Math.Log((FreqView[index] / (double)length));
-                Atomic.Add(ref entropy[new Index1D(0)], val);
-            }
-        }
-
-        static void IndexedCalcJointEntropyKernel(
-            Index1D index,
-            ArrayView2D<double, Stride2D.DenseX> FreqView,
-            ArrayView1D<double, Stride1D.Dense> FirstView,
-            ArrayView1D<double, Stride1D.Dense> SecondView,
-            ArrayView1D<double, Stride1D.Dense> entropy,
-            int length
-        )
-        {
-            double val;
-            Index2D ind2 = new Index2D((int)FirstView[index], (int)SecondView[index]);
-            if (FreqView[ind2] > 0)
-            {
-                val = -1 * (FreqView[ind2] / (double)length) * Math.Log((FreqView[ind2] / (double)length));
-                val = val / FreqView[ind2];
-                Atomic.Add(ref entropy[new Index1D(0)], val);
-            }
-        }
-
-        static void IndexedCalcJointEntropyKernel2(
-            Index1D index,
-            ArrayView<double> FreqView,
-            ArrayView<double> FirstView,
-            ArrayView<double> SecondView,
-            ArrayView<double> entropy,
-            int length,
-            long minSzudzikPair
-        )
-        {
-            double val;
-            var ind2 = szudzikPair((int)FirstView[index] - minSzudzikPair, (int)SecondView[index] - minSzudzikPair);
-            if (FreqView[ind2] > 0)
-            {
-                val = -1 * (FreqView[ind2] / (double)length) * Math.Log((FreqView[ind2] / (double)length));
-                val = val / FreqView[ind2];
-                Atomic.Add(ref entropy[new Index1D(0)], val);
-            }
-        }
-
-        static void IndexedCalcJointEntropyKernel3(
-            Index1D index,
-            ArrayView2D<double, Stride2D.DenseX> FreqView,
-            ArrayView1D<double, Stride1D.Dense> FirstView,
-            ArrayView1D<double, Stride1D.Dense> SecondView,
-            ArrayView1D<double, Stride1D.Dense> entropy,
-            int length
-        )
-        {
-            double val;
-            Index2D ind2 = new Index2D((int)FirstView[index], (int)SecondView[index]);
-            if (FreqView[ind2] > 0)
-            {
-                val = -1 * (FreqView[ind2] / (double)length) * Math.Log((FreqView[ind2] / (double)length));
-                val = val / FreqView[ind2];
-                Atomic.Add(ref entropy[new Index1D(0)], val);
-            }
-        }
-
-        static void normalizeArrayKernel(
-            Index1D index,
-            ArrayView1D<double, Stride1D.Dense> inputView,
-            ArrayView1D<double, Stride1D.Dense> outputView,
-            ArrayView1D<int, Stride1D.Dense> minVal
-        )
-        {
-            outputView[index] = Math.Floor(inputView[index]) - (double)minVal[new Index1D(0)];
-        }
-
-        static void GetMaxMinValKernal(
-            Index1D index,
-            ArrayView1D<double, Stride1D.Dense> aView,
-            ArrayView1D<int, Stride1D.Dense> MaxVal,
-            ArrayView1D<int, Stride1D.Dense> MinVal
-        )
-        {
-            if (aView[index] != Double.NaN)
-            {
-                Atomic.Max(ref MaxVal[new Index1D(0)], (int)Math.Floor(aView[index]));
-                Atomic.Min(ref MinVal[new Index1D(0)], (int)Math.Floor(aView[index]));
-                //Atomic.Add(ref MinVal[new Index1D(0)], 1);
-            }
-        }
-
-        static void TestGetMaxMinValKernal(
-            Index1D index,
-            ArrayView1D<double, Stride1D.Dense> aView,
-            ArrayView1D<int, Stride1D.Dense> MaxVal,
-            ArrayView1D<int, Stride1D.Dense> MinVal
-        )
-        {
-            if (!Double.IsNaN(aView[index]))
-            {
-                Atomic.Max(ref MaxVal[new Index1D(0)], (int)Math.Floor(aView[index]));
-                Atomic.Min(ref MinVal[new Index1D(0)], (int)Math.Floor(aView[index]));
-                //Atomic.Add(ref MinVal[new Index1D(0)], 1);
-            }
-        }
-
-        static void InsertBufferIntoTree(
+        protected static void InsertBufferIntoTree(
             Index1D index,
             ArrayView1D<int, Stride1D.Dense> View,
             ArrayView1D<int, Stride1D.Dense> Tree
         ) { }
 
-        static void InitMaxMinKernel(
-            Index1D index,
-            ArrayView1D<double, Stride1D.Dense> aView,
-            ArrayView1D<int, Stride1D.Dense> MaxVal,
-            ArrayView1D<int, Stride1D.Dense> MinVal
-        )
-        {
-            MaxVal[index] = (int)Math.Floor(aView[new Index1D(0)]);
-            MinVal[index] = (int)Math.Floor(aView[new Index1D(0)]);
-        }
-
-        ///<summary>Sets every element in buff to setvalue</summary>
-        ///<param name="buff">buff</param>
-        ///<param name="setvalue">setvalue</param>
-        static void setBuffToValueKernal(Index1D index, ArrayView1D<int, Stride1D.Dense> buff, int setvalue)
-        {
-            buff[index] = setvalue;
-        }
-
-        ///<summary>Sets every element in buff to setvalue</summary>
-        ///<param name="buff">buff</param>
-        ///<param name="setvalue">setvalue</param>
-        static void setBuffToValueDoubleKernal(Index1D index, ArrayView1D<double, Stride1D.Dense> buff, double setvalue)
-        {
-            buff[index] = setvalue;
-        }
-
-        ///<summary>Sets every element in buff to setvalue</summary>
-        ///<param name="buff">buff</param>
-        ///<param name="setvalue">setvalue</param>
-        static void setBuffToValue2DKernal(Index2D index, ArrayView2D<double, Stride2D.DenseX> buff, double setvalue)
-        {
-            buff[index] = setvalue;
-        }
-
-        static void BuildFreqKernel(
+        protected static void BuildFreqKernel(
             Index2D index,
             ArrayView1D<double, Stride1D.Dense> input,
             ArrayView1D<double, Stride1D.Dense> output
@@ -2290,21 +323,7 @@ namespace CSMI
             }
         }
 
-        static void BuildFreqAdjKernel(
-            Index1D index,
-            ArrayView1D<double, Stride1D.Dense> input,
-            ArrayView1D<double, Stride1D.Dense> output
-        )
-        {
-            if (!(Double.IsNaN(input[index])))
-            {
-                Atomic.Add(ref output[(int)Math.Floor(input[index])], 1.0);
-            }
-            //if(Math.Floor(input[index.X]) == Math.Floor(input[index.Y])){
-            //}
-        }
-
-        static void LargeBuildFreqKernel(
+        protected static void LargeBuildFreqKernel(
             Index3D index,
             ArrayView1D<double, Stride1D.Dense> input,
             ArrayView1D<double, Stride1D.Dense> output,
@@ -2321,75 +340,12 @@ namespace CSMI
             }
         }
 
-        static void BuildFreqAdjustedKernel(
-            Index1D index,
-            ArrayView1D<double, Stride1D.Dense> input,
-            ArrayView1D<double, Stride1D.Dense> output
-        )
-        {
-            //if(Math.Floor(input[index.X]) == Math.Floor(input[index.Y]) && Math.Floor(input[index.X]) != 0){
-            Atomic.Add(ref output[new Index1D((int)(input[index]))], 1.0);
-            //}
-        }
-
-        static void BuildJointFreqKernel(
-            Index1D index,
-            ArrayView1D<double, Stride1D.Dense> first,
-            ArrayView1D<double, Stride1D.Dense> second,
-            ArrayView2D<double, Stride2D.DenseX> output
-        )
-        {
-            if (!(Double.IsNaN(first[index])) && !(Double.IsNaN(second[index])))
-            {
-                //if(Math.Floor(first[index.X]) == Math.Floor(first[index.Y]) && Math.Floor(second[index.X]) == Math.Floor(second[index.Y])){
-                Atomic.Add(ref output[new Index2D((int)first[index], (int)second[index])], 1.0);
-            }
-            //}
-            // if(Math.Floor(first[index.X]) == Math.Floor(first[index.Y])){
-            //     Atomic.Add(ref firstmap[index.X], 1.0);
-            // }
-
-            // if(Math.Floor(second[index.X]) == Math.Floor(second[index.Y])){
-            //     Atomic.Add(ref secondmap[index.X], 1.0);
-            // }
-        }
-
-        static void BuildJointFreqKernel2(
-            Index1D index,
-            ArrayView<double> first,
-            ArrayView<double> second,
-            ArrayView<double> output,
-            long minSzudzikPair
-        )
-        {
-            if (!(Double.IsNaN(first[index])) && !(Double.IsNaN(second[index])))
-            {
-                Atomic.Add(
-                    ref output[(szudzikPair((int)first[index] - minSzudzikPair, (int)second[index] - minSzudzikPair))],
-                    1.0
-                );
-            }
-        }
-
-        static void BuildJointFreqKernel3(
-            Index1D index,
-            ArrayView1D<double, Stride1D.Dense> first,
-            ArrayView1D<double, Stride1D.Dense> second,
-            ArrayView2D<double, Stride2D.DenseX> output
-        )
-        {
-            if (!(Double.IsNaN(first[index])) && !(Double.IsNaN(second[index])))
-            {
-                Atomic.Add(ref output[new Index2D((int)first[index], (int)second[index])], 1.0);
-            }
-        }
-
-        void print1d<T>(T[] array)
+        public void print1d<T>(T[] array)
         {
             Console.WriteLine($"[{string.Join(", ", array)}]");
         }
 
-        void print2d<T>(T[,] array)
+        public void print2d<T>(T[,] array)
         {
             Console.WriteLine("[");
             for (int i = 0; i < array.GetLength(0); i++)
@@ -2465,19 +421,21 @@ namespace CSMI
             };
             // csharpier-ignore-start
             var resultsList = new double[]{
-                Utils.MeasureExecutionTime(javaResultsMap[0].name, () => calculateEntropy(a), printOutput: true),
-                Utils.MeasureExecutionTime(javaResultsMap[1].name, () => calculateConditionalEntropy(a, b), printOutput: true),
-                Utils.MeasureExecutionTime(javaResultsMap[2].name, () => calculateJointEntropy(a, b), printOutput: true),
-                // Utils.MeasureExecutionTime(javaResultsMap[2].name, () => calculateJointEntropy2(a, b), printOutput: true),
-                // Utils.MeasureExecutionTime(javaResultsMap[2].name, () => calculateJointEntropy3(a, b), printOutput: true), // TODO: finish or remove
-                Utils.MeasureExecutionTime(javaResultsMap[3].name, () => calculateMutualInformation(a, b), printOutput: true),
-                Utils.MeasureExecutionTime(javaResultsMap[4].name, () => calculateConditionalMutualInformation(a, b, c), printOutput: true),
+                Utils.MeasureExecutionTime(javaResultsMap[0].name, () => entropy.calculateEntropy(a), printOutput: true),
+                Utils.MeasureExecutionTime(javaResultsMap[1].name, () => entropy.calculateConditionalEntropy(a, b), printOutput: true),
+                Utils.MeasureExecutionTime(javaResultsMap[2].name, () => entropy.calculateJointEntropy(a, b), printOutput: true),
+                // Utils.MeasureExecutionTime(javaResultsMap[2].name, () => entropy.calculateJointEntropy2(a, b), printOutput: true),
+                // Utils.MeasureExecutionTime(javaResultsMap[2].name, () => entropy.calculateJointEntropy3(a, b), printOutput: true), // TODO: finish or remove
+                Utils.MeasureExecutionTime(javaResultsMap[3].name, () => mi.calculateMutualInformation(a, b), printOutput: true),
+                Utils.MeasureExecutionTime(javaResultsMap[4].name, () => mi.calculateConditionalMutualInformation(a, b, c), printOutput: true),
             };
             // csharpier-ignore-end
 
             // See if the functions for the C# vs Java implementation are the same
             var decimalPlaces = 12;
-            Console.WriteLine($"\nJava vs C# return values test results, accurate to '{decimalPlaces}' decimal places:");
+            Console.WriteLine(
+                $"\nJava vs C# return values test results, accurate to '{decimalPlaces}' decimal places:"
+            );
             for (int i = 0; i < resultsList.Length; i++)
             {
                 var csharpResult = Math.Round(resultsList[i], decimalPlaces);
@@ -2498,13 +456,13 @@ namespace CSMI
             double[] b = GenerateRandomNumbers(length);
             double[] c = GenerateRandomNumbers(length);
             // csharpier-ignore-start
-            Utils.MeasureExecutionTime("Calculate Entropy", () => calculateEntropy(a));
-            Utils.MeasureExecutionTime("Calculate Conditional Entropy", () => calculateConditionalEntropy(a, b));
-            Utils.MeasureExecutionTime("Calculate Joint Entropy", () => calculateJointEntropy(a, b));
-            // Utils.MeasureExecutionTime("Calculate Joint Entropy 2", () => calculateJointEntropy2(a, b));
-            // Utils.MeasureExecutionTime("Calculate Joint Entropy 3", () => calculateJointEntropy3(a, b)); // TODO: finish or remove
-            Utils.MeasureExecutionTime("Mutual Information", () => calculateMutualInformation(a, b));
-            Utils.MeasureExecutionTime("Conditional Mutual Information", () => calculateConditionalMutualInformation(a, b, c));
+            Utils.MeasureExecutionTime("Calculate Entropy", () => entropy.calculateEntropy(a));
+            Utils.MeasureExecutionTime("Calculate Conditional Entropy", () => entropy.calculateConditionalEntropy(a, b));
+            Utils.MeasureExecutionTime("Calculate Joint Entropy", () => entropy.calculateJointEntropy(a, b));
+            // Utils.MeasureExecutionTime("Calculate Joint Entropy 2", () => entropy.calculateJointEntropy2(a, b));
+            // Utils.MeasureExecutionTime("Calculate Joint Entropy 3", () => entropy.calculateJointEntropy3(a, b)); // TODO: finish or remove
+            Utils.MeasureExecutionTime("Mutual Information", () => mi.calculateMutualInformation(a, b));
+            Utils.MeasureExecutionTime("Conditional Mutual Information", () => mi.calculateConditionalMutualInformation(a, b, c));
             // csharpier-ignore-end
 
             Console.WriteLine();
@@ -2533,7 +491,7 @@ namespace CSMI
             }
             #endregion
 
-            using MI m = new MI();
+            using CSMI m = new CSMI();
 
             if (option_selected == 1)
             {
